% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{BioLink}
\alias{BioLink}
\title{Build the biological link}
\usage{
BioLink(PID, Mode="GO",ChemCas="default", ChemInchikey="default",
   DiseaseID="default", MetabolomeID="default", MetBiospec="blood", ProteomeID="default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Mode}{chr. Method to build the biological link between exposures and diseases.
Available options include "PPI" (i.e., protein-protein interaction) and "GO" (i.e., gene ontoloty).}

\item{ChemCas}{chr. CAS Registry Number of chemicals. Default means using the values in the input data file.
Users can also copy the part of them by clicking "Available vars". It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., "7440-43-9,333-41-5,20461-54-5".}

\item{ChemInchikey}{chr. InChiKey serial number of chemicals. Default means using the values in the input data file.
It should be noted that there is fixed format for the entering characters separated with comma and without space,
e.g., "WABPQHHGFIMREM-UHFFFAOYSA-N,BTAGRXWGMYTPBY-UHFFFAOYSA-N,IAKOZHOLGAGEJT-UHFFFAOYSA-N".}

\item{DiseaseID}{chr. ID of the concerned diseases. Both IDs from OMIM (e.g., OMIM:220100) and
MESH (e.g., MESH:C536409) are accepted. It should be noted that there is fixed format
for the entering characters separated with comma and without space,
e.g., "OMIM:244200,MESH:C536409,OMIM:181500".}

\item{MetabolomeID}{chr. KEGG entry number of metabolites. Default means using the values
in the input data file. Users can also copy the part of them by clicking "Available vars".
It should be noted that there is fixed format for the entering characters separated with comma
and without space, e.g., "C00022,C00117,C00794".}

\item{MetBiospec}{chr. Biological sample matrix for the metabolome analysis.
Options include "Blood" and "Urine".}

\item{ProteomeID}{chr. Protein ID. Both IDs of Ensembl and UniProt are accepted.
Default means using the values in the input data file. Users can also copy the part of them
by clicking "Available vars". It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "Q9Y3X0,Q8N5I3,ENSP00000000233,ENSP00000000412".}
}
\value{
A list object containing the edges and nodes of the biological link.
}
\description{
Build the biological link between the exposures and diseases
}
\examples{
res = InitBioLink()
   res1 = LoadBioLink(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
   res3 = BioLink(PID = res$PID, OutPath="default", Mode = "GO", ChemCas = "default",
   ChemInchikey = "default",DiseaseID = "default",MetabolomeID = "default",
   MetBiospec = "blood", ProteomeID = "default")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
