% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{ConvToExpoID}
\alias{ConvToExpoID}
\title{Convert different IDs to the unified ExposomeX IDs}
\usage{
ConvToExpoID(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}
}
\value{
A data frame containing the converted ID information.
}
\description{
Convert the IDs of exposure, chemicals, metabolites, or proteins to the unified ExposomeX ID,
i.e., unified identifier in ExposomeX platform
}
\examples{
res = InitBioLink()
   res1 = LoadBioLink(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
}
\author{
Mingliang Fang, Weinan Lin, Bin Wang (corresponding author)
}
