% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{DescComp}
\alias{DescComp}
\title{Size comparison for groups.}
\usage{
DescComp(PID = res$PID, Task = "mean", Vars = "X5,X6,X7,X8,X9", VarsBy = "Y1" , Method = "wilcox",
   Layout = "density" , Brightness = "dark" , Palette = "default3").
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{OutPath}{chr. Output file directory, e.g. "D:/StatDesc/DescComp". It should be
noted that the slash symbol is "/", not "\".If "default", the current working directory will be set.}

\item{Task}{chr. Comparison task. At present, only the "mean" comparison is available.}

\item{Vars}{chr. Target variables used for modelling. It should be noted that there is fixed format for the entering
characters separated with  "," and without space, e.g. "X5,X6,X7,X8,X9".
The default values is "all" (all variables are included).}

\item{VarsBy}{chr. Variable  used to group the observation for size comparison.}

\item{Method}{chr. Comparison method. At present, only "wilcox" (Wilcoxon rank sum test) is available.}

\item{Layout}{chr. Visualization layout. Available values include "column.points", "density".}

\item{Brightness}{chr. Visualization brightness. Available values include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available values include "default1", "default2" and 6 journal option
including "cell", "nature", "science", "lancet", "nejm" and "jama".}
}
\value{
A list object containing the results of size comparisons between groups for variables and  visualization of the results.
}
\description{
Perform size comparisons between groups.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatDesc(PID = res$PID)
   res3 = DescComp(PID=res$PID, OutPath="default",Task = "mean", Vars = "X5,X6,X7,X8,X9", VarsBy = "Y1",
   Method = "wilcox", Layout = "column.points",Brightness = "light" , Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
