% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{DescExtre}
\alias{DescExtre}
\title{Extreme value calculation.}
\usage{
DescExtre(PID=res$PID, Vars = "X5,X6,X7,X8,X9", LimitLow = 0.025, LimitUpper = 0.975,
   Layout = "column.points", Brightness = "light", Palette = "default2").
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{OutPath}{chr. Output file directory, e.g. "D:/StatDesc/DescExtre". It should be
noted that the slash symbol is "/", not "\".If "default", the current working directory will be set.}

\item{Vars}{chr. Target variables used for modelling. It should be noted that there is
fixed format for the entering characters separated with  "," and without space.
The default values is "all" (all variables are included).}

\item{LimitLow}{num. Lower limit ratio to screen the small extreme values located from 0 to this lower limit of the target variables.}

\item{LimitUpper}{num. Upper limit ratio to screen the large extreme values located from this lower limit to 1 of the target variables.}

\item{Layout}{chr. Visualization layout . Available values include "column.points", "heatmap".}

\item{Brightness}{chr. Visualization brightness . Available values include "light" and "dark".}

\item{Palette}{chr. Visualization palette . Available values include "default1", "default2" and 6 journal option
including "cell", "nature", "science", "lancet", "nejm" and "jama".}
}
\value{
A list object containing the results of extremum for variables and  visualization of the results.
}
\description{
Extreme value calculation for numeric variables.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatDesc(PID = res$PID)
   res3 = DescExtre(PID = res$PID, Vars = "X5,X6,X7,X8,X9", LimitLow = 0.025, LimitUpper = 0.975,
   Layout = "column.points", Brightness = "light", Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
