% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{DescNorm}
\alias{DescNorm}
\title{Normality test.}
\usage{
DescNorm(PID = res$PID, Vars = 'X5,X6,X7,X8,X9', Method = "shapiro.test", Layout = "rose.chart" ,
   Brightness = "light", Palette = "default3").
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{OutPath}{chr. Output file directory, e.g. "D:/StatDesc/DescNorm". It should be
noted that the slash symbol is "/", not "\".If "default", the current working directory will be set.}

\item{Vars}{chr. Target variables used for modelling. It should be noted that there is fixed format for
the entering characters separated with  "," and without space, e.g. "X7,X8,X9,X10,X11". The default
values is "all" (all variables are included).}

\item{Method}{chr. Normality test method. Only "shapiro.test" method is available at present.}

\item{Layout}{chr. Visualization layout. Available values include "column", "column.points", "rose.chart", and "density".}

\item{Brightness}{chr. Visualization brightness.  Available values include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available values include "default1", "default2" and 6 journal option
including "cell", "nature", "science", "lancet", "nejm" and "jama".}
}
\value{
A list object containing the results of normality test for variables and  visualization of the results.
}
\description{
Normality test for numeric variables.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatDesc(PID = res$PID)
   res3 = DescNorm(PID=res$PID, Vars = 'X5,X6,X7,X8,X9', Method = "shapiro.test", Layout = "column" ,
   Brightness = "light", Palette = "default2")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
