% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{DescSize}
\alias{DescSize}
\title{Variable description.}
\usage{
DescSize(PID=res$PID, Vars = "C1,C2,X5,X6,X7,X8,X9", VarsBy = "Y1", Layout = "box" , Brightness = "dark" ,
   Palette ="science").
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{OutPath}{chr. Output file directory, e.g. "D:/StatDesc/StatDesc". It should be
noted that the slash symbol is "/", not "\".If "default", the current working directory will be set.}

\item{Vars}{chr. Target variables used for modelling. It should be noted that there is fixed format for the entering characters
separated with  "," and without space, e.g. "X4,X5,X6,X7,X8". The default values is "all" (all variables are included).}

\item{VarsBy}{chr. Variable used to group the observation for size description.}

\item{Layout}{chr. Visualization layout. Available values include "box", "violin".}

\item{Brightness}{chr. Visualization brightness. Available values include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available values include "default1", "default2" and 5 journal option
including "jama", "nature", "science", "lancet", "nejm".}
}
\value{
A list object containing the results of variable description for continuous and discrete variables respectively
and  visualization of the continuous variables.
}
\description{
Calculate median, mean .etc for numeric variables,frequency and count for Discrete variable.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatDesc(PID = res$PID)
   res3 = DescSize(PID = res$PID, Vars = "X5,X6,X7,X8,X9", VarsBy = "Y1", Layout = "box",
   Brightness = "light", Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
