% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{DescTable1}
\alias{DescTable1}
\title{Create Table 1 for StatDesc Module.}
\usage{
DescTable1(PID = res$PID,EpiDesign = "cohort", VarsY = "Y1", VarsC = "C1,C2,C3,C4,C5,C6",
   Missing = "ifany").
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{OutPath}{chr. Output file directory, e.g. "D:/StatDesc/DescTable1". It should be
noted that the slash symbol is "/", not "\".If "default", the current working directory will be set.}

\item{EpiDesign}{chr. Research types provided for users, include "cohort" , "case.control", "cross.sectional".}

\item{VarsY}{chr. Outcome variable used for modelling. Only one variable can be entered.}

\item{VarsC}{chr. Covariate variables needing further statistical test.  It should be noted that there is
fixed format for the entering characters separated with  "," and without space, e.g. "C1,C2,C3,C4,C5".
The defaults value is all covariate variables listed in the data file, which can be entered with "all.c".}

\item{Missing}{chr. Counts of missing values in the table, available options include are "no" (never display missing values),
"ifany" (only display if any missing values), and "always" (includes missing count row for all variables). Default is "ifany".}
}
\value{
A list object containing standardized table 1
}
\description{
Create Table 1 for different epidemilogical study designs.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   res2 =  TidyStatDesc(PID = res$PID)
   res3 = DescTable1(PID = res$PID, EpiDesign = "cohort", VarsY = "Y1", VarsC = "C1,C2,C3,C4,C5,C6", Missing = "ifany")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
