% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EBIOBiolink}
\alias{EBIOBiolink}
\title{Build the biological link}
\usage{
EBIOBiolink(PID, OutPath="default",Mode, MetBiospec="blood")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEBIO.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Mode}{chr. Method to build the biological link between exposures and diseases.
Available options include "EPPD" (protein-protein interaction), "EGoD" (gene ontoloty),
"EPD" (protein), "EGeD" (Gene), "EPaD" (pathway).}

\item{MetBiospec}{chr. Biological sample matrix for the metabolome analysis.
Options include "blood" and "urine".}
}
\value{
A list object containing the edges and nodes of the biological link.
}
\description{
Build the biological link between the exposures and diseases
}
\examples{
res = InitEBIO()
   res1 = LoadEBIO(PID = res$PID,UseExample = "example#1")
   res2 = EBIOConvToExpoID(PID = res$PID)
   res3 = EBIOBiolink(PID = res$PID, OutPath="default", Mode = "EGoD", MetBiospec = "blood")
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Tianxiang Wu, Bin Wang,(corresponding author)
}
