% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EBIOConvToExpoID}
\alias{EBIOConvToExpoID}
\title{Convert different IDs to the unified ExposomeX IDs}
\usage{
EBIOConvToExpoID(PID,OutPath)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEBIO.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}
}
\value{
A data frame containing the converted ID information.
}
\description{
Convert the IDs of exposure, chemicals, metabolites, or proteins to the unified ExposomeX ID,
i.e., unified identifier in ExposomeX platform
}
\examples{
res = InitEBIO()
   res1 = LoadEBIO(PID = res$PID,UseExample = "example#1")
   res2 = EBIOConvToExpoID(PID = res$PID,OutPath = "default")
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Tianxiang Wu, Bin Wang,(corresponding author)
}
