% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EDBLink}
\alias{EDBLink}
\title{Find the nexuses between keywords}
\usage{
EDBLink(PID, OutPath = "default", ClassA, ClassB, LinkFrom, LinkTo)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{ClassA}{chr. Find the Links between the keywords in ClassA and ClassB.
Options include "Exposure","Protein","GO","Pathway","Gene"}

\item{ClassB}{chr. Options include "Protein","GO","Disease","Pathway","Gene"}

\item{LinkFrom}{chr. Name, alias, and ID of exposure, protein, GO, gene and pathway are all included.
"Default" means using the values in the input data file.}

\item{LinkTo}{chr. Name, alias, and ID of protein, GO, gene, pathwayand disease are all included.
"Default" means using the values in the input data file.}
}
\value{
A data frame
}
\description{
Find the nexuses between the keywords in ExposomeX platform.
Nexus direction from LinkFrom (class A) to LinkTo (class B)
}
\examples{
res <- InitEDB()
   res1 = LoadEDB(PID = res$PID, UseExample = "example#1")
   res2 = EDBLink(PID=res$PID,OutPath ="default",ClassA = "Exposure",
       ClassB = "Protein",LinkFrom = "default",LinkTo = "default")
   FuncExit(PID = res$PID)
}
\author{
Tianxiang Wu, Bin Wang (corresponding author)
}
