% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EDBNode}
\alias{EDBNode}
\title{Explain keywords}
\usage{
EDBNode(PID, OutPath = "default", Class, Nodes)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Class}{chr. Choose the search range of the concerned keywords. Options include
“Exposure” (In addition to containing chemicals, some exposure factors may be a mixture of chemicals, e.g. PM2.5, tobacco smoking.),
"Metabolite" (the chemicals used for metabolome analysis in the KEGG database),
"Protein", "Enzyme" (referring in particular to the enzymes in the KEGG database),
"Disease", "GO" (gene ontology), and “Gene".}

\item{Nodes}{chr. Any keywords to search. "default" means using the values in the
input data file. Users can also enter the keywords here. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., "7440-43-9,OMIM:619217,ARF5,GO:0010942".}
}
\value{
A data frame
}
\description{
Explain the keyword in ExposomeX platform
}
\examples{
res <- InitEDB()
   res1 = LoadEDB(PID = res$PID, UseExample = "example#1")
   res2 = EDBNode(PID=res$PID,OutPath = "default",Class = "GO",Nodes = "default")
   FuncExit(PID = res$PID)
}
\author{
Tianxiang Wu, Bin Wang (corresponding author)
}
