% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EVIZCateDot}
\alias{EVIZCateDot}
\title{Plot category dot}
\usage{
EVIZCateDot(PID,OutPath,Vars, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEVIZ.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Vars}{chr. Specifying the variables.
Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").
No more than 50 variables be entered is recommended (< 50 variables).}

\item{Parameter}{chr. Specifying which parameter of the data to be the ordinate of
the output plot. Available options include: "mean", "median", "min", "max", "mad"
or "sd".Default is "mean".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing the dot plot.
"light_default1": the visualization result.
}
\description{
Visualize data via dot plot.
}
\details{
The dot plot is used to display the relative position of
two data points in the same time period, or compare the difference
between the two categorical variables.
}
\examples{
res = InitEVIZ()
    res1 = LoadEVIZ(PID = res$PID, UseExample = "example#1")
    res2 = EVIZCateDot(PID=res$PID,Vars="X4,X5,X6,X7,X8,X9,X10",Parameter="mean",Brightness="light",Palette="default1")
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
