% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EVIZCompoDendrogram}
\alias{EVIZCompoDendrogram}
\title{Plot component dendrogram}
\usage{
EVIZCompoDendrogram(Vars, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEVIZ.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Vars}{chr. Variables to be visualized(e.g."X4,X5,X6,X7,X8,X9,X10").Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").}

\item{Parameter}{chr. Specifying which parameter of the data to be the ordinate of
the output plot. Available options include: "mean", "median", "min", "max", "mad"
or "sd".Default is "mean".}

\item{DistMethod}{chr.The distance measure. This must be one of "euclidean", "maximum"
or "manhattan".Default is "euclidean".}

\item{ClusterMethod}{chr.The agglomeration method. This should be one of "ward.D",
"ward.D2" or "single".Default is "ward.D".}

\item{ClusterNum}{num. The number of groups for cutting the tree.Default is 4.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing one plot.
(1) "light_default1": the whole dataset visualization result.
}
\description{
Visualize data via dendrogram plot.
}
\details{
The dendrogram plot is used to plot beautiful dendrograms.
}
\examples{
res = InitEVIZ()
    res1 = LoadEVIZ(PID = res$PID, UseExample = "example#1")
    res2= EVIZCompoDendrogram(PID=res$PID,Vars = "X4,X5,X6,X7,X8,X61,X66,X67,X200",Parameter = "median",DistMethod = "euclidean",ClusterMethod = "ward.D2",ClusterNum = "4",Brightness = "light",Palette = "default1")
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
