% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EVIZRelatMatrix}
\alias{EVIZRelatMatrix}
\title{Plot relationship matrix}
\usage{
EVIZRelatMatrix(PID,OutPath,Vars, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEVIZ.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Vars}{chr. Independent variables.Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").}

\item{Method}{chr. Method to calculate the correlation. Default option is "spearman".}
}
\value{
A list containing one plot.
(1) "light_default1": the whole dataset visualization result.
}
\description{
Visualize data via matrix plot.
}
\details{
The matrix plot is used to make a matrix of plots with a
given data set.
}
\examples{
res = InitEVIZ()
    res1 = LoadEVIZ(PID = res$PID, UseExample = "example#1")
    res2 = EVIZRelatMatrix(PID=res$PID,Vars = "X4,X5,X6,X7,X8,X9,X10",Method = "spearman")
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
