% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{EVIZRelatNetwork}
\alias{EVIZRelatNetwork}
\title{Plot relationship network}
\usage{
EVIZRelatNetwork(PID,OutPath,VarsY,VarsX, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEVIZ.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Specifying the dependent variables(e.g."Y2").}

\item{VarsX}{chr. Specifying the independent variables.Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").}

\item{Family}{chr. Specifying the data distribution in order to determine the visualization method.
Available options include:"gaussian","poisson" and "logistic".Notice that the family are
determined by data type of an outcome, or the plot can not be visualized.}

\item{Layout}{chr. Visualization layout. Available options include "force-directed" and "degree-circle".}

\item{CutOff}{num. Partial outcomes to visualize which is determined by correlation
coefficient r. The range must between 0 and 1.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing one plot.
(1) "light_default1": the visualization result;
}
\description{
Visualize data via network plot.
}
\details{
The network plot is used to visualize the relationship between
the input variables by using ggraph package.
}
\examples{
res = InitEVIZ()
    res1 = LoadEVIZ(PID = res$PID, UseExample = "example#1")
    res2 = EVIZRelatNetwork(PID=res$PID,VarsY="Y1",VarsX="all.x",Family="gaussian",Layout="force-directed",CutOff=0.8,Brightness="light",Palette="default1")
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
