% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expodb.R
\name{ExpoAbbr}
\alias{ExpoAbbr}
\title{Explain the abbreviations}
\usage{
ExpoAbbr(PID, OutPath = "default", Keys)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Keys}{chr. Any keywords to search. "default" means using the values .
in the input data file. Users can also enter the keywords here. It should be
noted that there is fixed format for the entering characters separated
with comma and without space, e.g., "EpiDesign,Cros,Cohort".}
}
\value{
A data frame
}
\description{
Explain the abbreviations in ExposomeX platform
}
\examples{
res <- InitDb()
   res1 = LoadDb(PID = res$PID, UseExample = "example#1")
   res2 = ExpoAbbr(PID=res$PID,OutPath = OutPath,Keys = "default")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
