% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expodb.R
\name{ExpoAnno}
\alias{ExpoAnno}
\title{Annotate the non-targeted features}
\usage{
ExpoAnno = function(PID, OutPath = "default", MassToCharge, AdductPos, 
    AdductNeg, Accuracy = 1)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{MassToCharge}{chr. Mass to charge ratio (m/z). It ranges 50-1000. If "default",
the values in the input data file are used. Users can also enter the keywords here.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "150,200,210".}

\item{AdductPos}{chr. Adducts collected in the positive mode. If "default", the values
in the input data file are chosen. Users can also enter the keywords here.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "M+H+Na,M+2ACN+2H,M+DMSO+H".
All the positive abducts are M+3ACN+2H,M+ACN+H,M+NH4,M+2ACN+2H,M+2H,M+3H,M+3Na,
"M+2Na-H,M+ACN+Na,M+H+Na,M+2K-H,M+H+NH4,2M+K,M+K,2M+NH4,2M+Na,M+2Na,M+DMSO+H,
M+2ACN+H,M+IsoProp+Na+H,M+2H+Na,M+ACN+2H,M+H,2M+H,M+CH3OH+H,M+H+2Na,M+Na",
2M+ACN+H,2M+ACN+Na,M+IsoProp+H,M+H+K"}

\item{AdductNeg}{chr. Adducts collected in the negative mode. If "default", the values
in the input data file are chosen. Users can also enter the keywords here.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "M+Hac-H,M-2H,M-H2O-H".
All the negative abducts are "M+FA-H,M+Hac-H,M+Br,3M-H,2M+Hac-H,M+K-2H,2M+FA-H,
M-H,M-H2O-H,M+Na-2H,M-2H,M+TFA-H,M+Cl,M-3H,2M-H".}

\item{Accuracy}{num. Upper limit of accuracy to match the target molecular.}
}
\value{
A data frame
}
\description{
Annotate the non-targeted features from high-resolution mass spectrometry
}
\examples{
res <- InitDb()
   res1 = LoadDb(PID = res$PID, UseExample = "example#1")
   res6 = ExpoAnno(PID=res$PID,OutPath = OutPath,MassToCharge = "default",
   AdductPos = "all",AdductNeg = "all",Accuracy = 5)
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
