% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expodb.R
\name{ExpoConv}
\alias{ExpoConv}
\title{Convert keywords}
\usage{
ExpoConv = function(PID, OutPath = "default", From, To, Keys)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{From}{chr. Choose the search range of the convert direction for the concerned keywords (from -> to).
Options include "chemical" <-> "cas.rn" (Chemical name <-> CAS Registry Number),
"inchikey" <-> "chemical" (InChIKey <-> Chemical name),
"metabolite" <-> "kegg.entry" (Metabolite name <-> InChIKey),
"kegg.entry" <-> "metabolite" (KEGG Entry ID <-> Metabolite name),
"protein" <-> "uniprot" (protein name <-> UniProt ID),
"uniprot" <-> "ensembl" (UniProt ID <-> Chemical name),
"enzyme" <-> "uniprot" (Enzyme name <-> UniProt ID),
"disease" <-> "disease.id" (Disease <-> Disease.id)}

\item{To}{chr. see "from".}

\item{Keys}{chr. Any keywords belong to the classes of "from" and "to" to search.
"default" means using the values in the input data file. Users can also enter
the keywords here. It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "7440-43-9,OMIM:619217,zinc,GO:0010942".}
}
\value{
A data frame
}
\description{
Convert the keywords from different sources in ExposomeX platform.
}
\examples{
res <- InitDb()
   res1 = LoadDb(PID = res$PID, UseExample = "example#1")
   res2 = ExpoConv(PID=res$PID,OutPath = OutPath,From = "chemical",To = "cas.rn",Keys = "default")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
