% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expodb.R
\name{ExpoDict}
\alias{ExpoDict}
\title{Explain keywords}
\usage{
ExpoDict(PID, OutPath = "default", Class, Keys)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoDB}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Class}{chr. Choose the search range of the concerned keywords. Options include
"chemical" (Some exposure factors may be a mixture of chemicals, e.g. PM2.5,
tobacco smoking. As chemicals account for the majority, we use "Chemical" for convenience),
"metabolite" (the chemicals used for metabolome analysis in the KEGG database),
"protein", "enzyme" (referring in particular to the enzymes in the KEGG database),
"disease", "GO" (gene ontology), and "ion.adduct" (the ion adducts in the
liquid chromatograph-mass spectrometry).}

\item{Keys}{chr. Any keywords to search. "default" means using the values in the
input data file. Users can also enter the keywords here. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., "7440-43-9,OMIM:619217,zinc,GO:0010942".}
}
\value{
A data frame
}
\description{
Explain the keyword in ExposomeX platform
}
\examples{
res <- InitDb()
   res1 = LoadDb(PID = res$PID, UseExample = "example#1")
   res2 = ExpoDict(PID=res$PID,OutPath = OutPath,Class = "GO",Keys = "default")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
