% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expocros.R
\name{FindCovaCros}
\alias{FindCovaCros}
\title{Find covariates}
\usage{
FindCovaCros = function(PID, OutPath = "default", VarsY, VarsC_Prior = "default", 
    VarsC_Fixed = NULL, Method = "single.factor", Thr = 0.1)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoCros}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsC_Prior}{chr. Potential covariates needing further statistical test.
The default value is all covariate variables listed in the data file.}

\item{VarsC_Fixed}{chr. Covariate variables fixed in the model by users.}

\item{Method}{chr. Methods for screening the covariates, including two options,
i.e. "single.factor" and "two.stage".}

\item{Thr}{num. Threshold of the P-value for screening the covariates. It is
ranging 0.05-0.25. The default value is 0.1.}
}
\value{
A list containing the selected covariates.
}
\description{
Find covariates
}
\examples{
res <- InitCros()
   res1 = LoadCros(PID = res$PID, UseExample = "example#1")
   res11 = FindCovaCros(PID=res$PID, VarsY = "Y1", 
   VarsC_Prior = "default", VarsC_Fixed = NULL, Method = "single.factor", Thr = 0.1)
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
