% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurv.R
\name{FindCovaSurv}
\alias{FindCovaSurv}
\title{Find covariates}
\usage{
FindCovaSurv = function(PID, OutPath = "default", TimeY, EventY, VarsC_Prior = "default", 
    VarsC_Fixed = NULL, Method = "single.factor", Thr = 0.1)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoCros}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{TimeY}{chr. Outcome variable of survival time used for modelling. Only one variable can be entered.}

\item{EventY}{chr. Outcome variable of status used for modelling. Only one variable can be entered.}

\item{VarsC_Prior}{chr. Potential covariates needing further statistical test.
The default value is all covariate variables listed in the data file.}

\item{VarsC_Fixed}{chr. Covariate variables fixed in the model by users.}

\item{Method}{chr. Methods for screening the covariates, including two options,
i.e. "single.factor" and "two.stage".}

\item{Thr}{num. Threshold of the P-value for screening the covariates. It is
ranging 0.05-0.25. The default value is 0.1.}
}
\value{
A list containing the selected covariates.
}
\description{
Find covariates
}
\examples{
res <- InitSurv()
   res1 = LoadSurv(PID = res$PID, UseExample = "example#1")
   res2 = TransDummy(PID=res$PID, Vars="default")
   res3 = FindCovaSurv(PID=res$PID, TimeY = "Y1", EventY= 'Y2',
   VarsC_Prior = "default", VarsC_Fixed = NULL, Method = "single.factor", Thr = 0.1)
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Bin Wang(corresponding author)
}
