% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It should be the same with the PID generated by Init.}
}
\value{
Exit status
}
\description{
End the module analysis
}
\examples{
res = InitEBIO()
   FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
