% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{HEPPredBlood}
\alias{HEPPredBlood}
\title{Build StatHEP model}
\usage{
HEPPredBlood(PID, OutPath, MC, N)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatHEP.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{MC}{lgl. T (or TRUE) and F (or FALSE). Whether to perform Mentocaro simulation.}

\item{N}{num. Times of Mentocaro simulation.}
}
\value{
A list containing StatHEP model results
}
\description{
HEPPredBlood function is designed to integrate the multi-omic data to predict the incidence risk.
It mainly aims to construct various stacked generalization models to predict the probability of outcome
incidence, as well as providing the statistical explanation.
}
\details{
Users can easily get the modeling results and their visualization plots with high quality by following the detailed instructions in each step.
Machine learning model is used to predict blood concentrations of chemicals and prioritize chemicals of health concern.
}
\examples{
res <- InitStatHEP()
   res1 <- LoadStatHEP(PID=res$PID, UseExample = "example#1", DataPath = NULL)
   res2 <- HEPPredBlood(PID=res$PID, OutPath = "default", MC = "T", N = 100)
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
