% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{InitBioLink}
\alias{InitBioLink}
\title{Initialize ExpoBioLink module}
\usage{
InitBioLink()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoBioLink module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
ExpoBioLink module is designed to find the biological relationships between
exposure factors and health outcome. This module adopts the most frequently-used and
authoritative databases, e.g., T3DB, CTD, ToxCast, StringDB, STITCH, KEGG, and GO.
}
\examples{
res <- InitBioLink()
}
\author{
Mingliang Fang, Bin Wang,(corresponding author)
}
