% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expocros.R
\name{InitCros}
\alias{InitCros}
\title{Initialize ExpoCros module}
\usage{
InitCros()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoCros module analysis. It can generate an R6 class object.
}
\details{
ExpoCros module was designed to analyze the cross-sectional data from
exposome-wide association study (EWAS). This data structure can be obtained
from the epidemiological designs of cross-section, case-control, and cohort.
}
\examples{
res <- InitCros()
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
