% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitEBIO}
\alias{InitEBIO}
\title{Initialize EBIO module}
\usage{
InitEBIO()
}
\value{
An R6 class object.
}
\description{
Initialize EBIO module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
EBIO module is designed to find the biological relationships between
exposure factors and health outcome. This module adopts the most frequently-used and
authoritative databases, e.g., T3DB, CTD, ToxCast, StringDB, STITCH, KEGG, and GO.
}
\examples{
res <- InitEBIO()
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Tianxiang Wu, Bin Wang,(corresponding author)
}
