% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitEDB}
\alias{InitEDB}
\title{Initialize EDB module}
\usage{
InitEDB()
}
\value{
An R6 class object.
}
\description{
Initialize EDB module analysis. It can generate an R6 class object.
}
\details{
EDB module is designed as a convenient tool to explore the data,
as well as facilitating to find the biological relationship between exposure
and diseases from the perspective of bioinformatics. This module adopts
the most frequently-used and authoritative databases,
e.g., T3DB, CTD, ToxCast, StringDB, STITCH, KEGG, and GO.
}
\examples{
res <- InitEDB()
    FuncExit(PID = res$PID)
}
\author{
Tianxiang Wu, Bin Wang (corresponding author)
}
