% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitEVIZ}
\alias{InitEVIZ}
\title{Initialize EVIZ module}
\usage{
InitEVIZ()
}
\value{
An R6 class object.
}
\description{
Initialize EVIZ module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
EVIZ module is designed for the data visualization of different
statistical and biological analyses in a user friendly and easy way, including
four typical classes of visualization.
}
\examples{
res <- InitEVIZ()
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
