% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomix.R
\name{InitMix}
\alias{InitMix}
\title{Initialize ExpoMixEffect module}
\usage{
InitMix()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoMixEffect module analysis. It can generate an R6 class object.
}
\details{
ExpoMixEffect module is designed to analyze mixture effect of the various
exposure factors. It mainly aims to screen the representative features with
high contribution to the health outcome, as well as their potential interaction effect.
}
\examples{
res <- InitMix()
}
\author{
Bin Wang (corresponding author)
}
