% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmo.R
\name{InitMO}
\alias{InitMO}
\title{Initialize ExpoMultiomics module}
\usage{
InitMO()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoMultiomics analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
ExpoMultiomics module is designed to integrate the multi-omic data to predict the incidence risk.
It mainly aims to construct various stacked generalization(SG) models to predict the probability
of outcome incidence, as well as providing the statistical explanation. In addition, the module can provide
visualization plots with high quality of the final calculation results to make it easier for users to understand.
}
\examples{
res <- InitMO()
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
