% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponta.R
\name{InitNTA}
\alias{InitNTA}
\title{Initialize ExpoNTA module}
\usage{
InitNTA()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoNTA module analysis. It can generate an R6 class object.
}
\details{
ExpoNontarget module is designed to conduct the analysis of the features
from the high-resolution mass spectrometry. It mainly aims to screen and
annotate the significant features associated with the health outcomes.
}
\examples{
res <- InitNTA()
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
