% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expopanel.R
\name{InitPanel}
\alias{InitPanel}
\title{Initialize ExpoPanel module}
\usage{
InitPanel()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoPanel module analysis. It can generate an R6 class object.
}
\details{
ExpoPanel module is designed to conduct the analysis of the panel data.
It mainly aims to evaluate the associations between exposure factors and the health outcome.
}
\examples{
res <- InitPanel()
}
\author{
Bin Wang (corresponding author)
}
