% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expostat.R
\name{InitStat}
\alias{InitStat}
\title{Initialize ExpoStat Module.}
\usage{
InitStat().
}
\value{
An R6 class object.
}
\description{
The first step to start ExpoStat Module.
}
\examples{
res = InitStat()
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
