% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatCros}
\alias{InitStatCros}
\title{Initialize StatCros module}
\usage{
InitStatCros()
}
\value{
An R6 class object.
}
\description{
Initialize StatCros module analysis. It can generate an R6 class object.
}
\details{
StatCros module was designed to analyze the cross-sectional data from
exposome-wide association study (EWAS). This data structure can be obtained
from the epidemiological designs of cross-section, case-control, and cohort.
}
\examples{
res <- InitStatCros()
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
