% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatDesc}
\alias{InitStatDesc}
\title{Initialize StatDesc Module.}
\usage{
InitStatDesc().
}
\value{
An R6 class object.
}
\description{
The first step to start StatDesc Module.
}
\examples{
res = InitStatDesc()
    FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
