% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatHEP}
\alias{InitStatHEP}
\title{Initialize StatHEP module}
\usage{
InitStatHEP()
}
\value{
An R6 class object.
}
\description{
Initialize StatHEP analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
Quantification of all trace organics in the biological fluids seems impossible and costly, regardless of the high individual exposure variability.
We hypothesized that the blood concentration of organic pollutants could be predicted via their exposure and chemical properties.
Developing a prediction model on the annotation of chemicals in human blood can provide new insight into the distribution and extent of exposures to a wide range of chemicals in humans.
Our objective of this module is to develop a machine learning model (eg., random forest) to predict blood concentrations of chemicals and prioritize chemicals of health concern.
Please see the details in Zhao et al., Environ Health Perspect. 2023 Mar;131(3):37009.
}
\examples{
res <- InitStatHEP()
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
