% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatIP}
\alias{InitStatIP}
\title{Initialize StatIP module}
\usage{
InitStatIP()
}
\value{
An R6 class object.
}
\description{
Initialize StatIP module analysis. It can generate an R6 class object.
}
\details{
StatIP module is designed to find the statistical relationships
between exposure factors and health outcome.
}
\examples{
res <- InitStatIP()
  FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
