% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expostatlink.R
\name{InitStatLink}
\alias{InitStatLink}
\title{Initialize ExpoStatLink module}
\usage{
InitStatLink()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoStatLink module analysis. It can generate an R6 class object.
}
\details{
ExpoStatLink module is designed to find the statistical relationships
between exposure factors and health outcome.
}
\examples{
res <- InitStatLink()
}
\author{
Bin Wang
}
