% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatMedt}
\alias{InitStatMedt}
\title{Initialize StatMedt module}
\usage{
InitStatMedt()
}
\value{
An R6 class object.
}
\description{
Initialize StatMedt module analysis. It can generate an R6
class object integrating all the analysis information.
}
\details{
InitStatMedt uses R6 package to generate an R6 class object where
parameters to be used for the following mediation module program are
initialized and save in that object. All executed function codes in the
ExpoMediaiton packaged will be recorded in the form of log text in that
object. Furthermore, a program ID (i.e., PID) is randomly created for the
users to identify their own program.
}
\examples{
res <- InitStatMedt()
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
