% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatMix}
\alias{InitStatMix}
\title{Initialize StatMix module}
\usage{
InitStatMix()
}
\value{
An R6 class object.
}
\description{
Initialize StatMix module analysis. It can generate an R6 class object.
}
\details{
StatMix module is designed to analyze mixture effect of the various
exposure factors. It mainly aims to screen the representative features with
high contribution to the health outcome, as well as their potential interaction effect.
}
\examples{
res <- InitStatMix()
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
