% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatPanel}
\alias{InitStatPanel}
\title{Initialize StatPanel module}
\usage{
InitStatPanel()
}
\value{
An R6 class object.
}
\description{
Initialize StatPanel module analysis. It can generate an R6 class object.
}
\details{
StatPanel module is designed to conduct the analysis of the panel data.
It mainly aims to evaluate the associations between exposure factors and the health outcome.
}
\examples{
res <- InitStatPanel()
    FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
