% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatSurv}
\alias{InitStatSurv}
\title{Initialize StatSurv module}
\usage{
InitStatSurv()
}
\value{
An R6 class object.
}
\description{
Initialize StatSurv module analysis. It can generate an R6
class object integrating all the analysis information.
}
\details{
InitStatSurv uses R6 package to generate an R6 class object where
parameters to be used for the following mediation module program are
initialized and save in that object. All executed function codes in the
StatSurv packaged will be recorded in the form of log text in that
object. Furthermore, a program ID (i.e., PID) is randomly created for the
users to identify their own program.
}
\examples{
res <- InitStatSurv()
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Bin Wang(corresponding author)
}
