% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{InitStatTidy}
\alias{InitStatTidy}
\title{Initialize StatTidy module}
\usage{
InitStatTidy()
}
\value{
An R6 class object.
}
\description{
Initialize StatTidy module analysis. It can generate an R6 class object.
}
\details{
It is designed to tidy the data for the target model analysis.
}
\examples{
res <- InitStatTidy()
     FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
