% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expotidy.R
\name{InitTidy}
\alias{InitTidy}
\title{Initialize ExpoTidy module}
\usage{
InitTidy()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoTidy module analysis. It can generate an R6 class object.
}
\details{
It is designed to tidy the data for the target model analysis.
}
\examples{
res <- InitTidy()
}
\author{
Bin Wang
}
