% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoviz.R
\name{InitViz}
\alias{InitViz}
\title{Initialize ExpoViz module}
\usage{
InitViz()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoViz module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
ExpoViz module is designed for the data visualization of different
statistical and biological analyses in a user friendly and easy way, including
four typical classes of visualization.
}
\examples{
res <- InitViz()
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
