% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{LoadBioLink}
\alias{LoadBioLink}
\title{Load data file for BioLink module}
\usage{
LoadBioLink(PID, UseExample = "default", DataPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{UseExample}{chr. Method of uploading data. If "default"，user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for BioLink module.
}
\examples{
res <- InitBioLink()
   res = LoadBioLink(PID = res$PID, UseExample = "example#1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
