% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expocros.R
\name{LoadCros}
\alias{LoadCros}
\title{Load data file for ExpoCros module}
\usage{
LoadCros(PID, UseExample = "default", DataPath=NULL, VocaPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoCros}

\item{UseExample}{chr. Method of uploading data. If "default"，user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_expocros.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input directory of vocabulary file, e.g. "D:/test/eg_voca_expocros.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for ExpoCros module
}
\examples{
res <- InitCros()
   res1 = LoadCros(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
