% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expodb.R
\name{LoadDb}
\alias{LoadDb}
\title{Load data file for ExpoDB module}
\usage{
LoadDb(PID, UseExample = "default", DataPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoDB}

\item{UseExample}{chr. Method of uploading data. If "default"，user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for ExpoDB module
}
\examples{
res <- InitDb()
   res1 = LoadDb(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
