% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadEBIO}
\alias{LoadEBIO}
\title{Load data file for EBIO module}
\usage{
LoadEBIO(PID, UseExample = "default", DataPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEBIO.}

\item{UseExample}{chr. Method of uploading data. If "default", user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for EBIO module.
}
\examples{
res <- InitEBIO()
   res = LoadEBIO(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Tianxiang Wu, Bin Wang,(corresponding author)
}
