% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadEMETA}
\alias{LoadEMETA}
\title{Load Data for EMETA Module}
\usage{
LoadEMETA(PID = NULL, UseExample = "default", DataPath = NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEMETA.}

\item{UseExample}{chr. A character indicates whether uses example data for analyses, available option include "example#1" for using example data1 and "default" for using data uploaded.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_meta.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Upload local data file for EMETA Module.
}
\details{
After initializing EMETA module, the second step is to upload local data file for EMETA Module.
Set param "UseExample = 'example#1'" to use example data1.
LoadEMETA can only run successfully after successfully running InitEMETA.
Please attention, PID must be got from the return result of InitEMETA().
}
\examples{
res = InitEMETA()
   res1 = LoadEMETA(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
