% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadEMS}
\alias{LoadEMS}
\title{Load data file for EMS module}
\usage{
LoadEMS(PID, UseExample = "default", MS2, DataPath=NULL, VocaPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EMS}

\item{UseExample}{chr. Method of uploading data. If "default", user should upload their own data files,
or use "example#1","example#2" provided by this module.}

\item{MS2}{lgl.Whether to use MS2 information to annotate the target ion. If use example data,
choose "T" to use "example#2" data with MS2 information, choose "F" to use "example#1" data
without MS2 information.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input directory of vocabulary file, e.g. "D:/test/eg_voca_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for EMS module
}
\examples{
res <- InitEMS()
   res1 = LoadEMS(PID = res$PID, UseExample = "example#1", MS2 = FALSE)
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
