% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponta.R
\name{LoadNTA}
\alias{LoadNTA}
\title{Load data file for ExpoNTA module}
\usage{
LoadNTA(PID, UseExample = "default", DataPath=NULL, VocaPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoNTA}

\item{UseExample}{chr. Method of uploading data. If "default"，user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input directory of vocabulary file, e.g. "D:/test/eg_voca_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for ExpoNTA module
}
\examples{
res <- InitNTA()
   res = LoadNTA(PID = res$PID, UseExample = "example#1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
