% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expostat.R
\name{LoadStat}
\alias{LoadStat}
\title{Load data for ExpoStat Module.}
\usage{
LoadStat(PID, UseExample = "default", DataPath = NULL, VocaPath = NULL).
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStat.}

\item{UseExample}{chr. Whether uses example data for analyses, available option include "example#1"
for using example data1 and "default" for using data.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/expostat_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input file directory, e.g. "D:/test/expostat_voca.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing imported data.
}
\description{
Upload  data file for ExpoStat Module.
}
\examples{
res = InitStat()
   res1 = LoadStat(PID = res$PID, UseExample = "example#1")
   res2 = LoadStat(PID = res$PID, UseExample = "default", DataPath = "D:/test/expostat_data.xlsx", VocaPath = "D:/test/expostat_voca.xlsx")
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
