% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadStatDesc}
\alias{LoadStatDesc}
\title{Load data for StatDesc Module.}
\usage{
LoadStatDesc(PID, UseExample = "default", DataPath = NULL, VocaPath = NULL).
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatDesc.}

\item{UseExample}{chr. Whether uses example data for analyses, available option include "example#1"
for using example data1 and "default" for using data.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/StatDesc_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input file directory, e.g. "D:/test/StatDesc_voca.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing imported data.
}
\description{
Upload  data file for StatDesc Module.
}
\examples{
res = InitStatDesc()
   res1 = LoadStatDesc(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Yanqiu Feng, Bin Wang (corresponding author)
}
