% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadStatHEP}
\alias{LoadStatHEP}
\title{Load data file for StatHEP module}
\usage{
LoadStatHEP(PID, UseExample= "default", DataPath)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatHEP.}

\item{UseExample}{chr. Method of uploading data. If "default", user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Upload data file for StatHEP module.
}
\examples{
res <- InitStatHEP()
    res <- LoadStatHEP(PID=res$PID, UseExample = "example#1", DataPath = NULL)
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
