% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadStatMix}
\alias{LoadStatMix}
\title{Load data file for StatMix module}
\usage{
LoadStatMix(PID, UseExample = "default", DataPath = NULL, VocaPath = NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatMix}

\item{UseExample}{chr. Method of uploading data. If "default",user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input directory of vocabulary file, e.g. "D:/test/eg_voca_biolink.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for StatMix module
}
\examples{
res <- InitStatMix()
   res = LoadStatMix(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
