% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadStatMO}
\alias{LoadStatMO}
\title{Load data file for multiomics module}
\usage{
LoadStatMO(PID, UseExample= "default", DataPath, VocaPath)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatMO.}

\item{UseExample}{chr. Method of uploading data. If "default", user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input file vocabulary, e.g. "D:/test/eg_voca.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Upload data file for multiomics module.
}
\examples{
res <- InitStatMO()
    res1 <- LoadStatMO(PID=res$PID, UseExample = "example#1", DataPath = NULL, VocaPath = NULL)
    FuncExit(PID = res$PID)
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
