% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{LoadStatSurv}
\alias{LoadStatSurv}
\title{Load data file for Survival module}
\usage{
LoadStatSurv(PID = PID, UseExample = "default", ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by
InitStatSurv.}

\item{UseExample}{chr. Method of uploading data. If "default",
user should upload their own data files, or use "example#1"
provided by this module.}

\item{DataPath}{chr. Input data file directory, e.g.
"D:/test/eg_Surv_data.xlsx". It should be noted that the slash
symbol is "/", not "\".}

\item{VocaPath}{chr. Input vocabulary file directory, e.g.
"D:/test/eg_Surv_voca.xlsx". It should be noted that the slash
symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data and vocabulary file.
}
\description{
Load data file for Survival module
}
\details{
LoadStatSurv function loads the data file and the vocabulary file into
the R6 object that InitStatSurv created. Noted that there are several data
format requirments for the data and vocabulary file.
For data file,the first three columns should be named as "SampleID", "SubjectID", and "Group", respectively.
For the "Group" variable, only two values can be used, i.e. "train" and "test". If there is no data for test, all values should be set as "train".
For outcome variables, their initials must be set as "Y" and serialized by adding Arabic numerals if needed, e.g., Y1, Y2, Y3. In this module, the survival time (Y1) and status (Y2) must be provided.For exposure variables, their initials must be set as "X" and serialized by adding Arabic numerals if needed, e.g., X1, X2, X3.
For covariate variables, their initials must be set as "C" and serialized by adding Arabic numerals if needed, e.g., C1, C2, C3. It should be noted the covariates are not required if users don’t have.
For vocabulary file, the first two columns must be named as "SerialNo" and "FullName", respectively.
The list of SerialNo of outcomes, exposure, and covariates should be the same with the column names of "Data file".
The list of the FullName is prepared as users’ like.
}
\examples{
res <- InitStatSurv()
   res1 <- LoadStatSurv(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Bin Wang(corresponding author)
}
