% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{MetaRef}
\alias{MetaRef}
\title{Search or Download Articles' Main Information}
\usage{
MetaRef(PID = NULL, Mode = "search", VarX, VarY, VarM, YearFrom, YearEnd, PMID, OutPath)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEMETA.}

\item{Mode}{chr. Two modes are provided. “Search” for paper retrieval by keywords VarX/VarY/VarM/YearFrom/YearEnd, and “Download” for downloading main informain (main information only) for specified PMID.}

\item{VarX}{chr. “default” or a chemical names character (separate different values by ","). If “default”, the function will use the VarX values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the chemical names in the character instead.}

\item{VarY}{chr. “default” or a disease names character (separate different values by ","). If “default”, the function will use the VarY values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the disease names in the character instead.}

\item{VarM}{chr. “default” or a mediating factor names character (separate different values by ","). If “default”, the function will use the VarM values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the mediating factor names in the character instead.}

\item{YearFrom}{chr. “default” or a disease names character (separate different values by ","). Limits the time range searched for "search" mode. If “default”, the function will use the YearFrom values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the YearFrom in the character instead.}

\item{YearEnd}{chr. “default” or a disease names character (separate different values by ","). Limits the time range searched for "search" mode. If “default”, the function will use the YearEnd values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the YearEnd in the character instead.}

\item{PMID}{chr. “default” or a disease names character (separate different values by ","). Refs to the papers PMID for "download" mode. If “default”, the function will use the PMID values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the PMIDs in the character instead.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default", the current working directory will be set. It should be noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing dataframe of articles' information.
}
\description{
Search or download articles' main information based on keywords.
}
\details{
MetaRef provides the functions of paper retrieval and relevance sorting, returning the information to the user based on keywords.
Please attention, PID must be got from the return result of InitEMETA().
MetaRef can only run successfully after successfully running InitEMETA and LoadEMETA functions.
}
\examples{
res = InitEMETA()
   res1 = LoadEMETA(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
   res2 = MetaRef(PID = res$PID,
   OutPath = "default",
   Mode = "search",
   VarX = "default",
   VarY = "default",
   VarM = "default",
   YearFrom = "default",
   YearEnd = "default",
   PMID = "default")
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
