% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{MetaRev}
\alias{MetaRev}
\title{Review Relationship Between Exposure and Outcome}
\usage{
MetaRev(PID = NULL, CID = "default", DID = "default", OutPath = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEMETA.}

\item{CID}{chr. “default” or a chemical ID character (separate different values by ","). If “default”, the function will use the Chemical_ID values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the chemical ID in the character instead. Chemical_ID refers to the target chemical ID which can be inchikey (eg. JIAARYAFYJHUJI-UHFFFAOYSA-L ), cas.rn (eg. 7784-42-1) or our EXC ID (eg. EX:C01631).}

\item{DID}{chr. “default” or a disease ID character (separate different values by ","). If “default”, the function will use the Disease_ID values in the file loaded by LoadEMETA. If a character (separate different values by ","), the function will use the disease ID in the character instead. Disease_ID refers to the target disease ID which can be MESH ID (format like MESH:D006973), OMIM ID (format like OMIM:182940) or our EXD ID (eg. EX:D16243)}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default", the current working directory will be set. It should be noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing relationship visualization.
}
\description{
Review the relationship between exposure and outcome besed on chemical ID and disease ID.
}
\details{
MetaRev provides the functions of literature review. In the publishd papers, how many recorded that X is a protective/risky factor for Y? This questions can be solved by MetaRev function. (It can only search the papers available in our database)
Please attention, PID must be got from the return result of InitEMETA().
MetaRev can only run successfully after successfully running InitEMETA and LoadEMETA functions.
}
\examples{
res = InitEMETA()
   res1 = LoadEMETA(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
   res2 = MetaRev(PID = res$PID,
   OutPath = "default",
   CID = "default",
   DID = "default")
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
