% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{MixBKMR}
\alias{MixBKMR}
\title{Build the Bayesian Kernel Machine Regression (BKMR) model}
\usage{
MixBKMR(PID, OutPath = "default", VarsY, VarsX, IncCova,
    Family, Group = F,Iter = 2000, qfixed = 0.5, qsbivar = "default",
    qsoverall = "default", qsdiff = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatMix}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{Covariates}{chr. To determine covariates in models. Default is "all.c", where all
candidate covariates will be included in models. Users can also choose available
variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "C1,C2,C3".}

\item{Family}{chr. The link function for the regression model according the data type of outcomes,
including "gaussian" for continuous variable, "binomial" for binary variable,
and "poisson" for counting variable}

\item{Group}{lgl. T (or TRUE) and F (or FALSE). Whether to use group indicators
for fitting hierarchical variable selection. If "TRUE", the group name (GroupName)
should be provided in the vocabulary data file.}

\item{Iter}{num. Number of iterations for modeling. The default is 500.
For more accurate and stable results, a minimum of 10,000 iteration is recommended.}

\item{qfixed}{num. Quantile at which to fix the other predictors. The default is 0.5.}

\item{qsbivar}{chr. Quantiles at which to fix the second variable. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., the default character sequence is "0.1,0.5,0.9"}

\item{qsoverall}{num. Quantiles at which to calculate the overall risk summary. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., the default character sequence is "0.25,0.30,0.35,0.40,0.45,0.50,0.55,0.60,0.65,0.70,0.75"}

\item{qsdiff}{chr. Indicating the two quantiles for computing their effect difference. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., the default character sequence is "0.25,0.75".}
}
\value{
A list containing the BKMR analysis results.
}
\description{
Build the Bayesian Kernel Machine Regression (BKMR) model
}
\examples{
res <- InitStatMix()
   res1 = LoadStatMix(PID = res$PID, UseExample = "example#1")
   res2 <- TidyStatMix(PID = res$PID, LogTransM = "log10", RangeLow = 0, RangeUpp = 1)
   res3 = MixBKMR(PID = res$PID, VarsY = "Y1", VarsX = "X4,X5,X6,X7,X8,X9,X10",
   Covariates = "all.c", Family = "gaussian", Group = 'F', Iter = 2000, qfixed = 0.5,
   qsbivar = "default", qsoverall = "default", qsdiff = "default")
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
