% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{MixGcomp}
\alias{MixGcomp}
\title{Build g-computation model}
\usage{
MixGcomp(PID, VarsY, VarsX, Covariates = "all.c", Family, q = 10)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatMix}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{Covariates}{chr. To determine covariates in models. Default is "all.c", where all
candidate covariates will be included in models. Users can also choose available
variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "C1,C2,C3".}

\item{Family}{chr. The link function for the regression model according the data type of outcomes,
including "gaussian" for continuous variable, "binomial" for binary variable,
and "poisson" for counting variable.}

\item{q}{num. Levels for ranking mixture variables, e.g. in quartiles (q = 4), deciles (q = 10),
or percentiles (q = 100). The default is 10.}
}
\value{
A list containing the WQS analysis results.
}
\description{
Build g-computation model
}
\examples{
res <- InitStatMix()
   res1 = LoadStatMix(PID = res$PID, UseExample = "example#1")
   res2 <- TidyStatMix(PID = res$PID, LogTransM = "log10", RangeLow = 0, RangeUpp = 1)
   res3 = MixGcomp(PID=res$PID, VarsY = "Y1", VarsX = "all.x",
   Covariates = "all.c", Family = "gaussian", q = 10)
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
