% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{MixMLR}
\alias{MixMLR}
\title{Build multiple linear regression (MLR) model}
\usage{
MixMLR(PID,OutPath = "default",VarsY,VarsX,IncCova = "F",
    SelMethod = "lasso",PredType = "response",Family)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatMix}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available
variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{Covariates}{chr. To determine covariates in models. Default is "all.c", where all
candidate covariates will be included in models. Users can also choose available
variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "C1,C2,C3".}

\item{SelMethod}{chr. Method to select the important features to the final model.
Options include "stepwise" (stepwise regression), "lasso" (Regularization regression of
least absolute shrinkage and selection operator), and "enet" (Regularization regression of elastic net).}

\item{PredType}{chr. Prediction type of the outcome variable, including "response"
for the actual values and "prob" for outcome with binary variable.}

\item{Family}{chr. The link function for the regression model according the data type of outcomes,
including "gaussian" for continuous variable, "binomial" for binary variable,
and "poisson" for counting variable}
}
\value{
A list containing the MLR analysis.
}
\description{
Build multiple linear regression (MLR) model
}
\examples{
res <- InitStatMix()
   res1 = LoadStatMix(PID = res$PID, UseExample = "example#1")
   res2 <- TidyStatMix(PID = res$PID, LogTransM = "log10", RangeLow = 0, RangeUpp = 1)
   res3 = res3 = MixMLR(PID = res$PID, VarsY = "Y1", VarsX = "all.x",
   Covariates = "all.c", SelMethod = "lasso", PredType = "response",
   Family = "gaussian")
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
