% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{NtaAnno}
\alias{NtaAnno}
\title{Annotate the non-targeted features}
\usage{
NtaAnno(PID=res$PID,OutPath="default",VarsY = "Y1",VarsX = "default",VarsN = "single.factor",
   FdrCorrect = "F",AdductPos = "M+H",AdductNeg = "M-H",AccuracyMS1 = 5,MS2 = "F",
   AccuracyMS2 = 20,ScoreMS2 = 0.4,DiffRT = 5)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EMS}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{FdrCorrect}{lgl. T (or TRUE) and F (or FALSE). Whether to correct the
multiple hypotheses by false positive rate (FDR) method.}

\item{AdductPos}{chr. Adducts formed in the positive mode using LC-MS.
The default is "all", i.e., all the adducts in positive mode will be chosen, including
"M+3ACN+2H,M+ACN+H,M+NH4,M+2ACN+2H,M+2H,M+3H,M+3Na,M+2Na-H,M+ACN+Na,M+H+Na,
M+2K-H,M+H+NH4,2M+K,M+K,2M+NH4,2M+Na,M+2Na,M+DMSO+H,M+2ACN+H,M+IsoProp+Na+H,
M+2H+Na,M+ACN+2H,M+H,2M+H,M+CH3OH+H,M+H+2Na,M+Na,2M+ACN+H,2M+ACN+Na,M+IsoProp+H,M+H+K"}

\item{AdductNeg}{chr. Adducts formed in the negative mode using LC-MS.
The default is "all", i.e., all the adducts in positive mode will be chosen, including
"M+FA-H,M+Hac-H,M+Br,3M-H,2M+Hac-H,M+K-2H,2M+FA-H,M-H,M-H2O-H,M+Na-2H,
M-2H,M+TFA-H,M+Cl,M-3H,2M-H”}

\item{AccuracyMS1}{num. Accuracy threshold to match the target compounds of MS1 information.
The unit is ppm. It is usually set ranging 1-20.}

\item{MS2}{lgl. T (or TRUE) and F (or FALSE). Whether to use MS2 information to annotate the target ion.}

\item{AccuracyMS2}{num. Accuracy threshold to match the target compounds of MS2 information.
The unit is ppm. It is usually set ranging 1-20.}

\item{ScoreMS2}{num. The lowest matching score ranging 0-1 using MS2 information. The default is 0.4.}

\item{DiffRT}{num. Difference of the retention time between the measured and the theoretical prediction.
The unit is second (s). It is usually set ranging 60-120.}
}
\value{
A list containing non-target analysis results
}
\description{
Annotate the non-targeted features
}
\examples{
res <- InitEMS()
   res1 = LoadEMS(PID = res$PID, UseExample = "example#1", MS2 = FALSE)
   res2 = TidyEMS(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = NtaCros(PID=res$PID, OutPath="default",VarsY = "Y1", VarsX = "all.x",
   VarsN = "single.factor", FdrCorrect = "T", SelMethod = "lasso",StepwizeThr = 0.1,
   RF_ImpThr = 0.9, Covariates = "all.c", Family = "gaussian", RepMsr = "F", Corstr = "ar1")
   res4 = NtaAnno(PID=res$PID,OutPath="default",VarsY = "Y1",VarsX = "all.x",VarsN = "single.factor",
   FdrCorrect = "F",AdductPos = "M+H",AdductNeg = "M-H",AccuracyMS1 = 5,MS2 = "F",
   AccuracyMS2 = 20,ScoreMS2 = 0.4,DiffRT = 5)
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
