% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{PanelAsso}
\alias{PanelAsso}
\title{Association analysis of panel data}
\usage{
PanelAsso = function(PID, OutPath = "default", VarsY, VarsX, VarsN = "single.factor",
    VarsRandomIpt = "SubjectID", VarsRandomSlp = "none",Covariates = "all.c")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatPanel}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{VarsRandomIpt}{chr. Random intercept variable for the linear mixed-effect model.
The default is "SubjectID".}

\item{VarsRandomSlp}{chr. Random slope variable for the linear mixed-effect model.
The default is "none". It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{Covariates}{chr. Covariates used for modeling. The default option is
"all.c" (All covariates are included).}
}
\value{
A list containing the association analysis results.
}
\description{
Association analysis of panel data
}
\examples{
res <- InitStatPanel()
    res1 = LoadStatPanel(PID = res$PID, UseExample = "example#1")
    res2 = TidySataPanel(PID = res$PID, OutPath = "default",TransDummyVars="default")
    res3 = PanelAsso(PID=res$PID, OutPath = "default", VarsY = "Y1",
    VarsX = "X1,X2,X3,X4,X5,X6,X7,X8,X9,X10,X11,X12", VarsN = "single.factor",
    VarsRandomIpt = "SubjectID", VarsRandomSlp = "none", Covariates = "all.c")
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
