% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{StatIPCros}
\alias{StatIPCros}
\title{Build statistical link for cross-sectional data.}
\usage{
StatIPCros(PID, OutPath = "default", VarsY, VarsX, LinkModel = "ranger",
    ObsrPartType = "raw",ObsrPartNum = "50",ObsrProfType = "partial",
    ObsrProfNum = "100",ObsrProfVars = "all.x",ObsrProfGeom = "profiles",
    SubjPredSeq = "none",SubjPartType = "break_down")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatIP}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modelling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modelling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{LinkModel}{chr. Methods to interpret the model. Options include "ranger" (random forest),
"glmnet" (elastic net), "svm" (support vector machine), "glm" (linear regression),
"gam" (generalized additive model), and "xgboost" (eXtreme gradient boosting).}

\item{ObsrPartType}{chr. Type of transformation that should be applied for dropout loss.
Options include "raw" (drop losses), "ratio" (drop_loss/drop_loss_full_model),
and "difference" (drop_loss - drop_loss_full_model)}

\item{ObsrPartNum}{chr. Number of observations that should be sampled for calculation
of variable importance. The default means variable importance will be calculated
on whole dataset (no sampling). If "defult", use all Obsrrvations.}

\item{ObsrProfType}{chr. Type of variable profile. Options include "partial",
"conditional", and "accumulated"}

\item{ObsrProfNum}{int. Number of observations used for calculation of aggregated profiles.
By default 100.}

\item{ObsrProfVars}{chr. Names of variables to be explained. If "all.x", all "X variable" above are chosen.}

\item{ObsrProfGeom}{chr. Layout of the explanation profile in dataset level
including "aggregates", "profiles" or "points".}

\item{SubjPredSeq}{chr. Subjects which need explanation. Options include "all" (all the subjects),
"none" (no subjects), and "other" (copy the subject list by clicking "Available vars").}

\item{SubjPartType}{chr. Layout of the explanation profile in subject level.
Options include "shap", "oscillations", "break_down", and "none".}
}
\value{
A list containing all the statistical explanation results.
}
\description{
Build statistical link for cross-sectional data.
}
\examples{
res <- InitStatIP()
   res1 = LoadStatIP(PID = res$PID, UseExample = "example#1")
   res2 = StatIPCros(PID=res$PID, OutPath = "default", VarsY = "Y1" ,VarsX = "all.x",
   LinkModel = "ranger",ObsrPartType = "raw" ,ObsrPartNum  = "50",
   ObsrProfType = "partial" ,ObsrProfNum = "100", ObsrProfVars = "all.x",
   ObsrProfGeom = "profiles",SubjPredSeq = "S1,S2,S3",SubjPartType = "break_down")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
