% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{SurvAsso}
\alias{SurvAsso}
\title{Association analysis}
\usage{
SurvAsso = function(PID, OutPath = "default",
    TimeY, EventY,VarsX='all.x',VarsN = "single.factor",VarsSel = F,
    Covariates = "all.c")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatSurv()}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{TimeY}{chr. Outcome variable of survival time used for modelling. Only one variable can be entered.}

\item{EventY}{chr. Outcome variable of status used for modelling. Only one variable can be entered.}

\item{VarsX}{Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{VarsSel}{lgl. T (or TRUE) and F (or FALSE). Whether to select the significant variable for the final model.
Available options includes T and F}

\item{Covariates}{chr. Covariates used for modeling. The default option is
"all.c" (All covariates are included).}
}
\value{
A list containing the association analysis results.
}
\description{
Association analysis for survival data
}
\examples{
res <- InitStatSurv()
    res1 = LoadStatSurv(PID = res$PID, UseExample = "example#1")
    res2 = TidyStatSurv(PID = res$PID, OutPath = "default",TransDummyVars="default")
    res3 = SurvAsso(PID=res$PID, TimeY = "Y1", EventY= 'Y2',VarsX='all.x',
    VarsN="single.factor",VarsSel="T",Covariates = "all.c")
    FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Bin Wang(corresponding author)
}
