% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{SurvPred}
\alias{SurvPred}
\title{Build prediction models}
\usage{
SurvPred = function(PID,OutPath = "default",TimeY,EventY,VarsX = "all.x",
    Covariates = "all.c",RsmpMethod = "cv",Folds = 3,Ratio = 0.667,Repeats = 3)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatSurv()}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{TimeY}{chr. Outcome variable of survival time used for modelling. Only one variable can be entered.}

\item{EventY}{chr. Outcome variable of status used for modelling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{Covariates}{chr. Covariates used for modeling. The default option is
"all.c" (All covariates are included).}

\item{RsmpMethod}{chr. Three resampling methods options for internal validation, including
"cv" (i.e.,Cross validation) , "bootstrap", and "holdout".}

\item{Folds}{num. Folds of Cross-validation resampling. It is ranging 2-10.}

\item{Ratio}{num. Ratio of Bootstrap resampling. It is ranging 0.4-0.9.}

\item{Repeats}{num. Number of Bootstrap resampling. It is ranging 2-20.}
}
\value{
A list containing the prediction performance evaluation.
}
\description{
Build prediction models
}
\examples{
res <- InitStatSurv()
   res1 = LoadStatSurv(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatSurv(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = SurvPred(PID=res$PID, TimeY = "Y1", EventY= 'Y2',VarsX='all.x',
   Covariates = "all.c",RsmpMethod="cv",Folds=3,Ratio=0.667,Repeats=3)
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Bin Wang(corresponding author)
}
