% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{TidyDelMiss}
\alias{TidyDelMiss}
\title{Delete variables with missing values}
\usage{
TidyDelMiss(PID, OutPath = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}
}
\value{
An R6 class object containing the variable(s) without missing values.
}
\description{
Whether to delete missing variables with low variance. The default option is "yes".
If skipped, it may result in failure during modeling.
}
\examples{
res = InitStatTidy()
    res1 = LoadStatTidy(PID=res$PID, UseExample="example#1")
    res1$Expo$Data
    res2 = TidyDelMiss(PID=res$PID)
    res2$Expo$Data
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
