% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{TidyEMS}
\alias{TidyEMS}
\title{Tidy data.}
\usage{
TidyEMS(PID, OutPath = "default",TransDummyVars="default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EMS}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{TransDummyVars}{chr. Variables to be transformed as dummy variables. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., "X1,X2,X3". If "default", all the factor variables will be transformed into dummy ones.}
}
\value{
An R6 class object containing the variable(s) after transforming the factor
variables into dummy ones.
}
\description{
Tidy data file for EMS module
}
\examples{
res <- InitEMS()
    res1 = LoadEMS(PID = res$PID, UseExample = "example#1", MS2 = FALSE)
    res2 = TidyEMS(PID = res$PID, OutPath = "default",TransDummyVars="default")
    FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
