% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{TidyStatMix}
\alias{TidyStatMix}
\title{Tidy data for Mix module}
\usage{
TidyStatMix(PID = PID, TransDummyVars = ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by
InitStatMix.}

\item{TransDummyVars}{chr. A vector indicating factor variables to be transformed
as dummy variables. Default value is "default" where all factor vars will be
converted into dummies.}

\item{LogTransM}{chr. Method of converting data distribution. Available options
include "log10"(default), "log2", and "ln".}

\item{RangeLow}{num. Lower range value for data scaling. Default value is 0.
RangeLow value should be lower than that of RangeUpp.}

\item{RangeUpp}{num. Upper range value for data scaling. Default value is 1.
RangeUpp value should be larger than that of RangeLow.}
}
\value{
An R6 class object containing the input data and vocabulary file.
}
\description{
Tidy data file for Mix module
}
\details{
TidyStatMix function resembles several tidy steps for data pre-treatment
before executing other functions in Mix module, including deleting vars with
low variance, transforming categorical vars into factors, transferring data
distribution and scaling for exposures.
}
\examples{
res <- InitStatMix()
   res1 <- LoadStatMix(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- TidyStatMix(PID = res$PID, LogTransM = "log10", RangeLow = 0, RangeUpp = 1)
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
