% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{TidyTransType}
\alias{TidyTransType}
\title{Transform data type}
\usage{
TidyTransType(PID,OutPath = "default",Vars,To)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables;
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{To}{chr. Indicate the type of the chosen variables to be transformed into.
Available options include "integer", "numeric", "character", "factor", "logical", and "date".}
}
\value{
An R6 class object containing the variable(s) after transforming data type.
}
\description{
Transform data type
}
\examples{
res = InitStatTidy()
    res1 = LoadStatTidy(PID=res$PID, UseExample="example#1")
    res1$Expo$Data
    res2 = TidyTransType(PID=res$PID, Vars = "X1,X2", To = "character")
    res2$Expo$Data
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
