% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expotidy.R
\name{TransDistr}
\alias{TransDistr}
\title{Transform variable distribution}
\usage{
TransDistr(PID,OutPath = "default", Vars, Method)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables; "all.c", all covariates;
"all.cx", combination of All X and All C. Users can also choose available variables
by selecting "Other" option, and copy the variables by clicking "Available vars".
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{Method}{}
}
\value{
An R6 class object containing the variable(s) after transforming distribution.
}
\description{
Transform variable distribution
}
\examples{
res = InitTidy()
    res1 = LoadTidy(PID=res$PID, UseExample="example#1")
    res1$Expo$Data
    res2 = TransDistr(PID=res$PID, Vars="X6,X7", Method="log10")
    res2$Expo$Data
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
