% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expotidy.R
\name{TransDummy}
\alias{TransDummy}
\title{Transform factor variables into dummy ones}
\usage{
TransDummy(PID, Vars="default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{chr. Variables to be transformed as dummy variables. It should be noted that
there is fixed format for the entering characters separated with comma and without space,
e.g., "X1,X2,X3". If "default", all the factor variables will be transformed into dummy ones.
These variables need to be transformed as factor ones in previous transform step
using TransType function.}
}
\value{
An R6 class object containing the variable(s) after transforming the factor
variables into dummy ones.
}
\description{
Transform factor variables into dummy ones
}
\examples{
res = InitTidy()
    res1 = LoadTidy(PID=res$PID, UseExample="example#1")
    res1$Expo$Data
    res2 = TransDummy(PID=res$PID, Vars="default")
    res2$Expo$Data
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
