% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expotidy.R
\name{TransGroup}
\alias{TransGroup}
\title{Transform exposure groups}
\usage{
TransGroup = function(PID, Vars="default", ToGroup)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables; "all.c", all covariates;
"all.cx", combination of All X and All C. Users can also choose available variables
by selecting "Other" option, and copy the variables by clicking "Available vars".
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{ToGroup}{chr. Label the group of the target variables. Four common used names
are recommended, including Exposure, Metabolome, Proteome, and Immunome.
Users can also label the groups as you like.}
}
\value{
An R6 class object containing the variable(s) after grouping the variables.
}
\description{
Transform exposure groups
}
\examples{
res = InitTidy()
    res1 = LoadTidy(PID=res$PID, UseExample="example#1")
    res1$Expo$Data
    res2 = TransGroup(PID=res$PID, Vars="X4,X5", ToGroup = "G1")
    res2$Expo$Data
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
