% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{VizBioLink}
\alias{VizBioLink}
\title{Visualize the biological link}
\usage{
BioLink = function(PID,OutPath="default",Mode,ChemCas = "default",ChemInchikey = "default",DiseaseID = "default",
    MetabolomeID = "default",MetBiospec = "blood",ProteomeID = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Mode}{chr. Method to build the biological link between exposures and diseases.
Available options include "PPI" (i.e., protein-protein interaction) and "GO" (i.e., gene ontoloty).}

\item{Layout}{chr. Visualization layout. Available options include "force-directed" and "degree-circle".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list object containing the plot of the biological link. This plot can be further processed using ggplot2 package.
}
\description{
Visualize the biological link. It should be noted that the corresponding link
has been built by BioLink function prior to using it.
}
\examples{
res = InitBioLink()
   res1 = LoadBioLink(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
   res3 = BioLink(PID = res$PID, OutPath="default", Mode = "GO", ChemCas = "default",
   ChemInchikey = "default",DiseaseID = "default",MetabolomeID = "default",
   MetBiospec = "blood", ProteomeID = "default")
   res4 = VizBioLink(PID = res$PID, Mode = 'GO', Layout = "force-directed",
   Brightness = "dark", Palette = "default1")
}
\author{
Mingliang Fang, Ning Gao, Bin Wang (corresponding author)
}
