% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizCrosAsso}
\alias{VizCrosAsso}
\title{Visualize association analysis}
\usage{
VizCrosAsso = function(PID, OutPath = "default",VarsN = "single.factor",
    Layout = "volcano",Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatCros}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "volcano".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}

\item{ColorFor}{chr.
#' @param SizeFor chr}
}
\value{
An R6 class object containing the results' plot.
}
\description{
Visualize association analysis
}
\examples{
res <- InitStatCros()
   res1 = LoadStatCros(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatCros(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = CrosAsso(PID=res$PID, Linear = TRUE, EpiDesign = "cohort",
   VarsY = "Y1", VarsX = "X5,X6,X7,X8,X9,X10,X11", VarsN = "single.factor" ,
   VarsSel = FALSE, VarsSelThr = 0.1, Covariates = "all.c", Family = "gaussian",
   RepMsr = FALSE,Corstr = "ar1")
   res4 = VizCrosAsso(PID=res$PID, VarsY = "Y1", VarsN="single.factor", Layout = "volcano",
   Brightness = "dark",Palette = "default1", ColorFor = "beta.value", SizeFor = "p.value")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
