% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizCrosPred}
\alias{VizCrosPred}
\title{Visualize the prediction performance}
\usage{
VizCrosPred = function(PID,OutPath = "default",Layout = "bar",
    Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatCros}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{Layout}{chr. Visualization layout. Available options include "bar" and "roc"}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing the results' plot.
}
\description{
Visualize the prediction performance
}
\examples{
res <- InitStatCros()
   res1 = LoadStatCros(PID = res$PID, UseExample = "example#1")
   res3 = CrosPred(PID=res$PID, VarsY = "Y1", VarsX = "X5,X6,X7,X8,X9,X10,X11",
   PredType = "response",VarsSel = FALSE,VarsSelThr = 0.1, Covariates = "all.c",
   RsmpMethod = "cv" ,Folds = 5,Ratio = 0.667,Repeats = 5)
   res4 = VizCrosPred(PID=res$PID, VarsY = "Y1", Layout = "bar",
   Brightness = "light", Palette = "science")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
