% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoviz.R
\name{VizDistrSierra}
\alias{VizDistrSierra}
\title{Plot distribution sierra}
\usage{
VizDistrSierra(PID=res$PID,OutPath="default",Group = "F",Vars = "X14,X15,X16,X17,X18,X19,X20",Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Group}{lgl.Whether to separate dataset into train and test datasets for
data imputation,including T or F.The default option is F.}

\item{Vars}{chr. Variables to be visualized(e.g."X4,X5,X6,X7,X8,X9,X10").Available options include:
"all.x", all independent variables;
"all.c", all covariate variables;
"all.cx", combination of all.x and all.c;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").
No more than 50 variables be entered is recommended (< 20 variables).}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and "Journal". The "Journal" option provides several journal preference styles
including cell, nature, science, lancet, nejm, and jama.}
}
\value{
A list containing one or three plots.
When Group=T,
(1) "all_light_default1": the whole dataset visualization result;
(2) "train_light_default1": the train dataset visualization result;
(3) "test_light_default1": the test dataset visualization result.
When Group=F,
(1) "light_default1": the whole dataset visualization result.
}
\description{
Visualize data via sierra plot.
}
\details{
The sierra plot is used to visualize the kernel density
estimation of data.
}
\examples{
res = InitViz()
    res1 = LoadViz(PID = res$PID, UseExample = "example#1")
    res2 = VizDistrSierra(PID=res$PID,OutPath="default",Group = "F",
    Vars = "X14,X15,X16,X17,X18,X19,X20",Brightness = "light",
    Palette = "default1")
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
