% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizHEPPredBlood}
\alias{VizHEPPredBlood}
\title{Visualize StatHEP results}
\usage{
VizHEPPredBlood(PID, OutPath, MC,Layout,Brightness,Palette)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatHEP.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{MC}{lgl. T (or TRUE) and F (or FALSE). Whether to perform Mentocaro simulation.}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "boxplot".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2" and other options about some journal preference styles
including "cell", "nature", "science", "lancet", "nejm", etc.}
}
\value{
An R6 class object.
}
\description{
VizHEPPredBlood function is mainly aimed to visualize the modeling results calculated by HEPPredBlood function.
It can provide plots with high quality of the final results to make it easier for users to understand.
}
\details{
You can get different styles of images by selecting different parameters.
}
\examples{
res <- InitStatHEP()
   res1 <- LoadStatHEP(PID=res$PID, UseExample = "example#1", DataPath = NULL)
   res2 <- HEPPredBlood(PID=res$PID, OutPath = "default", MC = "T", N = 100)
   res3 <- VizHEPPredBlood(PID=res$PID, OutPath = "default", MC = "T",
   Layout = "forest", Brightness = "light", Palette = 'default1')
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Ning Gao, Bin Wang (corresponding author)
}
