% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizMedtPair}
\alias{VizMedtPair}
\title{Visualize pairwise mediation modelling result}
\usage{
VizMedtPair(PID, OutPath, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitStatMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}

\item{Brightness}{chr. Visualization brightness. Available options include
"light" and "dark".}

\item{Pallette}{chr. Visualization palette. Available options include "default1",
"default2" and "Journal". The "Journal" option provides several journal
preference styles including cell, nature, science, lancet, nejm, and jama.}
}
\value{
A list containing two elements where the visualized pairwise modelling plot
as well as the organized plotting data are stored. The elements of that list
include:
\enumerate{
\item "plotdata": a dataframe containing the organized plotting data extracted from #'     pairwise mediation modelling result.
\item "plot": a visualized plot for pairwise mediation result.
}
}
\description{
Visualize the pairwise mediation result. It should be noted that the
pairwise mediation modelling result has been built by Pairwise function
prior to using it.
}
\details{
VizMedtPair draws a visualized display for pairwise modelling results.
Piror to using VizMedtPair, make sure that the users have built mediation
models by Pairwise function.
}
\examples{
res <- InitStatMedt()
   res1 <- LoadStatMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 = TidyStatMedt(PID = res$PID, OutPath = "default",LogTransM = "log10",
   RangeLow = 0,RangeUpp = 1)
   res3 <- Pairwise(PID=res$PID, VarsY = "Y1",
   VarsX = "X1,X2,X3", VarsM = "M1,M2,M3", VarsC = "C1", Family = "linear",
   Iter = 500)
   res4 <- VizMedtPair(PID=res$PID, Brightness = "light",
   Pallette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
