% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizMultOmic}
\alias{VizMultOmic}
\title{Visualize multiomics model results}
\usage{
VizMultOmic(PID, OutPath, VarsY, NodeNum, EdgeThr, Layout, Brightness, Palette)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatMO.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable for modeling. Only one variable can be entered.}

\item{NodeNum}{num. Number of nodes in the network plot. The maximum number is generated by Multiomics function.
User can set a smaller value if needed.}

\item{EdgeThr}{num. Threshold of correlation coefficient ranging 0-1 for generating the concerned edges of the network plot.}

\item{Layout}{chr. Visualization layout. Available options include "force-directed" and "degree-circle".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2" and other options about some journal preference styles
including "cell", "nature", "science", "lancet", "nejm", etc.}
}
\value{
An R6 class object containing seven elements. The elements of that object include:
(1) "Importance_plot": Plots for the importance of features after modeling a single omic from different learners.
(2) "Measures_boxplot": Plots for the r-square value of the single omic model built by different learners.
(3) "NetWork_State": A list containing dataframes that contain nodes and edges used to draw
interOmic network from models built by different learners.
(4) "Nodeplot": Interomic node plots from models built by different models.
(5) "Networkplot": Interomic network plots from models built by different models.
(6) "Prediction_train_plot": Plots for the prediction value of the train set in the SG model built by different combinations of learners.
(7) "Prediction_test_plot": Plots for the prediction value of the test set in the SG model built by different combinations of learners.
}
\description{
VizMultOmic function is mainly aimed to visualize the modeling results calculated by MulOmicsCros function.
It can provide plots with high quality of the final results to make it easier for users to understand.
}
\details{
You can get different styles of images by selecting different parameters.
}
\examples{
res <- InitStatMO()
   res1 <- LoadStatMO(PID=res$PID, UseExample = "example#1", DataPath = NULL, VocaPath = NULL)
   res2 <- TidyStatMO(PID=res$PID, Vars = "X1,X2", To = "factor")
   res3 <- MulOmicsCros(PID=res$PID, OutPath = "default", OmicGroups = "immunome,metabolome",
   VarsY = "Y1", VarsC = "all.c", TuneMethod = "random_search", TuneNum = 2, RsmpMethod = "cv", Folds = 2,
   Ratio = 0.67, Repeats = 5, VarsImpThr = 0.85, SG_Lrns ="lasso,enet")
   res4 <- VizMultOmic(PID=res$PID, OutPath = "default", VarsY = "Y1", NodeNum=100, EdgeThr= 0.45,
   Layout = "force-directed", Brightness = "light", Palette = 'default1')
   FuncExit(PID = res$PID)
}
\author{
Guohuan Zhang, Yuting Wang, Ning Gao, Bin Wang (corresponding author)
}
