% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizNtaAnno}
\alias{VizNtaAnno}
\title{Visualize annotation results}
\usage{
VizNtaAnno(PID=res$PID,OutPath= "default",VarsY = "Y1",VarsN = "single.factor",
   AccuracyMS1 = 5,MS2 = "F",AccuracyMS2 = 20,DiffRT = 100,Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EMS}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{AccuracyMS1}{num. Accuracy threshold to match the target compounds of MS1 information.
The unit is ppm. It is usually set ranging 1-20.}

\item{MS2}{lgl. T (or TRUE) and F (or FALSE). Whether to use MS2 information to annotate the target ion.}

\item{AccuracyMS2}{num. Accuracy threshold to match the target compounds of MS2 information.
The unit is ppm. It is usually set ranging 1-20.}

\item{DiffRT}{num. Difference of the retention time between the measured and the theoretical prediction.
The unit is second (s). It is usually set ranging 60-120.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing plots of non-target annotation analysis results
}
\description{
Visualize annotation results of non-targeted data
}
\examples{
res <- InitEMS()
   res1 = LoadEMS(PID = res$PID, UseExample = "example#1", MS2 = "F")
   res2 = TidyEMS(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = NtaCros(PID=res$PID, OutPath="default",VarsY = "Y1", VarsX = "all.x",
   VarsN = "single.factor", FdrCorrect = "T", SelMethod = "lasso",
   StepwizeThr = 0.1, RF_ImpThr = 0.9, Covariates = "all.c", Family = "gaussian",
   RepMsr = "F", Corstr = "ar1")
   res4 = NtaAnno(PID=res$PID,OutPath="default",VarsY = "Y1",VarsX = "all.x",VarsN = "single.factor",
   FdrCorrect = "F",AdductPos = "M+H",AdductNeg = "M-H",AccuracyMS1 = 5,MS2 = "F",
   AccuracyMS2 = 20,ScoreMS2 = 0.4,DiffRT = 5)
   res5 = VizNtaAnno(PID=res$PID,OutPath="default",VarsY = "Y1",VarsN = "single.factor",
   AccuracyMS1 = 5,MS2 = "F",AccuracyMS2 = 20,DiffRT = 100,Brightness = "light",Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
