% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizNtaCros}
\alias{VizNtaCros}
\title{Visualize the results of Association analysis}
\usage{
VizNtaCros(PID, OutPath = "default", VarsY,VarsN, Layout = "volcano",
    EffectThr = 0.5, Brightness = "light", Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by EMS}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor".}

\item{Layout}{chr. Visualization layout. Available options include "volcano" and "forest".}

\item{EffectThr}{num. Threshold of the total importance for the variables to a
random forest model. It ranges 0.5-1.0 with the default value of 0.9.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing plots of non-target analysis association results
}
\description{
Visualize the results of Association analysis for non-targeted data
}
\examples{
res <- InitEMS()
   res1 = LoadEMS(PID = res$PID, UseExample = "example#1", MS2 = FALSE)
   res2 = TidyEMS(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = NtaCros(PID=res$PID, OutPath="default",VarsY = "Y1", VarsX = "all.x",
   VarsN = "single.factor", FdrCorrect = "T", SelMethod = "lasso",
   StepwizeThr = 0.1, RF_ImpThr = 0.9, Covariates = "all.c", Family = "gaussian",
   RepMsr = "F", Corstr = "ar1")
   res4 = VizNtaCros(PID=res$PID,OutPath="default",VarsY = "Y1",VarsN = "single.factor",
   Layout = "volcano",EffectThr = 0.5,Brightness = "light",Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
