% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizPanelAsso}
\alias{VizPanelAsso}
\title{Visualize the results of association analysis for panel data}
\usage{
VizPanelAsso = function(PID, OutPath, VarsY, VarsN, Layout = "volcano",
    Brightness = "dark", Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by StatPanel}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "volcano".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing the plots of association analysis results.
}
\description{
Visualize the results of association analysis for panel data
}
\examples{
res <- InitStatPanel()
    res1 = LoadStatPanel(PID = res$PID, UseExample = "example#1")
    res2 = TidySataPanel(PID = res$PID, OutPath = "default",TransDummyVars="default")
    res3 = PanelAsso(PID=res$PID, OutPath = "default", VarsY = "Y1",
    VarsX = "X1,X2,X3,X4,X5,X6,X7,X8,X9,X10,X11,X12", VarsN = "single.factor",
    VarsRandomIpt = "SubjectID", VarsRandomSlp = "none", Covariates = "all.c")
    res4 = VizPanelAsso(PID = res$PID, OutPath = "default", VarsY = "Y1",
    VarsN = "single.factor", Layout = "forest", Brightness = "dark",Palette = "default1")
    FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
