% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizRefer}
\alias{VizRefer}
\title{Visualize the Articles' Main Information}
\usage{
VizRefer(PID = NULL, OutPath = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEMETA.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default", the current working directory will be set. It should be noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing plots of article information visualization.
}
\description{
Visualize the articles' main information after MetaRef function.
}
\details{
VizRefer provides the visual function of articles' main information, showing the year and region distribution directly.
Please attention, PID must be got from the return result of InitEMETA().
VizRefer can only run successfully after successfully running InitEMETA, LoadEMETA and MetaRef functions.
}
\examples{
res = InitEMETA()
  res1 = LoadEMETA(PID = res$PID,
  UseExample = "example#1",
  DataPath = NULL)
  res2 = MetaRef(PID = res$PID,
  OutPath = "default",
  Mode = "search",
  VarX = "default",
  VarY = "default",
  VarM = "default",
  YearFrom = "default",
  YearEnd = "default",
  PMID = "default")
  res3 = VizRefer(PID=res$PID,
  OutPath = "default")
  FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
