% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exviz.R
\name{VizRelatEdgeBundling}
\alias{VizRelatEdgeBundling}
\title{Plot elationship edge bundling}
\usage{
VizRelatEdgeBundling(PID,OutPath="default",VarsY, VarsC="all.c", VarsX = "all.x",
    Family = "gaussian", SizeFor = "pvalue",Brightness = "light",
    Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitViz.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted
that the slash symbol is "/", not "\".If "default", the current working
directory will be set.}

\item{VarsY}{chr. Dependent variables for visualization(e.g."Y2").}

\item{VarsC}{chr. Covariate variable.Available options include:
"all.c", all covariate variables;
or input a character string specifying the variables,separated by comma ","
without space(e.g."C1,C2").}

\item{VarsX}{chr. Independent variables.Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma ","
without space(e.g."X4,X5,X6,X7,X8,X9,X10").}

\item{Family}{chr. The link function for the regression model according the data
type of outcomes,including "gaussian" for continuous variable, "binomial" for
binary variable,and "poisson" for counting variable. Notice that the family are
determined by data type of an outcome, or the plot can not be visualized.The
default option is "gaussian".}

\item{SizeFor}{chr. Parameter to represent the size of the points in the output plot.
Available options include "pvalue" and "beta".The default option is "pvalue".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1",
"default2" and "Journal". The "Journal" option provides several journal preference
styles including cell, nature, science, lancet, nejm, and jama.}
}
\value{
A ggplot plot which can be further modified using the ggplot2 package.
(1) "light_default1": the visualization result;
}
\description{
Visualize data via edge bundling plot.
}
\details{
The edge bundling plot is used to bundle the edges closely in order to
reduce complexity.
}
\examples{
res = InitViz()
    res1 = LoadViz(PID = res$PID, UseExample = "example#1")
    res2 = VizRelatEdgeBundling(PID=res$PID,OutPath="default",VarsY = "Y2",
    VarsX = "all.x",VarsC = "all.c",Family = "gaussian" ,SizeFor = "pvalue",
    Brightness = "light",Palette = "default1")
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
