% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoviz.R
\name{VizRelatHeatmap}
\alias{VizRelatHeatmap}
\title{Plot relationship heatmep}
\usage{
VizRelatHeatmap(PID,OutPath,VarsY,VarsX, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Group}{lgl.Whether to separate dataset into train and test datasets for d
ata imputation,including T or F.The default option is T.}

\item{VarsY}{chr. Dependent variables for visualization(e.g."Y2").}

\item{VarsX}{chr. Independent variables.Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma "," without space(e.g."X4,X5,X6,X7,X8,X9,X10").}

\item{Method}{chr. Method to calculate the correlation. Default option is "spearman".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and "Journal". The "Journal" option provides several journal preference styles
including cell, nature, science, lancet, nejm, and jama.}
}
\value{
A list containing one or two plots.
When Group=T,
(1) "train_light_default1": the train dataset visualization result;
(2) "test_light_default1": the test dataset visualization result.
When Group=F,
(1) "light_default1": the whole dataset visualization result.
}
\description{
Visualize data via heatmap plot.
}
\details{
The heatmap plot is used to display data in color changes
as a matrix.
}
\examples{
res = InitViz()
    res1 = LoadViz(PID = res$PID, UseExample = "example#1")
    res2 = VizRelatHeatmap(PID=res$PID,Group = "T",VarsY = "Y2",VarsX = "X1,X4,X5,X6,X7,X8,X9,X10",Method = "spearman",Brightness = "light",Palette = "default1")
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
