% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exviz.R
\name{VizRelatMatrix}
\alias{VizRelatMatrix}
\title{Plot relationship matrix}
\usage{
VizRelatMatrix(PID,OutPath="default",Group = "F",VarsY,VarsX,Method = "spearman")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitViz.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted t
hat the slash symbol is "/", not "\". If "default", the current working directory
will be set.}

\item{Group}{lgl.Whether to separate dataset into train and test datasets for
data imputation,including T or F.The default option is F.}

\item{VarsY}{chr. Dependent variables for visualization(e.g."Y2").}

\item{VarsX}{chr. Independent variables.Available options include:
"all.x", all independent variables;
or input a character string specifying the variables,separated by comma ","
without space(e.g."X4,X5,X6,X7,X8,X9,X10").
Notice that no more than 20 variables can be inputed.(< 20 variables).}

\item{Method}{chr. Method to calculate the correlation. Default option is "spearman".}
}
\value{
One or two ggmatrix plots which can be further modified using the ggplot2 package.
When Group=T,
(1) "Train": the train dataset visualization result;
(2) "Test": the test dataset visualization result.
When Group=F,
(1) "All": the whole dataset visualization result.
}
\description{
Visualize data via matrix plot.
}
\details{
The matrix plot is used to make a matrix of plots with a given data set.
}
\examples{
res = InitViz()
    res1 = LoadViz(PID = res$PID, UseExample = "example#1")
    res2 = VizRelatMatrix(PID=res$PID,OutPath="default",Group = "F",
    VarsY = "Y2",VarsX = "X4,X5,X6,X7,X8,X9,X10",Method = "spearman")
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
