% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizSurvCompGroup}
\alias{VizSurvCompGroup}
\title{Compare the survival curves of two groups}
\usage{
VizSurvCompGroup = function(PID,OutPath = "default",TimeY,EventY,
    VarsG,Model='km',VarsAdj,AdjMethod='average',Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatSurv()}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{TimeY}{chr. Outcome variable of survival time used for modelling. Only one variable can be entered.}

\item{EventY}{chr. Outcome variable of status used for modelling. Only one variable can be entered.}

\item{VarsG}{chr.  Grouping variable, must be a binary variable.}

\item{Model}{chr. Methods to depict the survival curve. Options include 'km' (Kaplan-Meier estimate) and
"coxph" (Cox proportional hazards regression mode).}

\item{VarsAdj}{If you choose the cox model, co-variables used for modelling.
It should be noted that there is fixed format for the entering characters separated with comma and without space, e.g., X1,X2,X3.}

\item{AdjMethod}{If you choose the cox model, method for adjusting model, include: "average","single","margin" and "conditional".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., "cell", "nature", "science", "lancet", "nejm", and "jama").}
}
\value{
A list containing the results' plot.
}
\description{
Compare the survival curves of two groups
}
\examples{
res <- InitStatSurv()
   res1 = LoadStatSurv(PID = res$PID, UseExample = "example#1")
   res2 = VizSurvCompGroup(PID=res$PID,TimeY="Y1",EventY="Y2",VarsG="C3",
   Model="km",Brightness="light",Palette='default1')
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Ning Gao, Bin Wang(corresponding author)
}
