% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomex.R
\name{VizSurvPred}
\alias{VizSurvPred}
\title{Visualize the prediction performance}
\usage{
VizSurvPred = function(PID,OutPath = "default",Layout = "curve",
    Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitStatSurv()}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Layout}{chr. Visualization layout. Available options include "curve" , "bar" and 'all'.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., "cell", "nature", "science", "lancet", "nejm", and "jama").}
}
\value{
A list containing the results' plot.
}
\description{
Visualize the prediction performance
}
\examples{
res <- InitStatSurv()
   res1 = LoadStatSurv(PID = res$PID, UseExample = "example#1")
   res2 = TidyStatSurv(PID = res$PID, OutPath = "default",TransDummyVars="default")
   res3 = SurvPred(PID=res$PID, TimeY = "Y1", EventY= 'Y2',VarsX='all.x',
   Covariates = "all.c",RsmpMethod="cv",Folds=3,Ratio=0.667,Repeats=3)
   res4 = VizSurvPred(PID=res$PID,Layout="curve",Brightness="light",Palette='default1')
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Ning Gao, Bin Wang(corresponding author)
}
