% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extidy.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It should be the same with the PID generated by initial functions.}
}
\value{
Exit status
}
\description{
End the module analysis
}
\examples{
res = InitTidy()
   res1 = LoadTidy(PID=res$PID, UseExample="example#1")
   res2 = DelNearZeroVar(PID=res$PID)
   FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
