% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoviz.R
\name{LoadViz}
\alias{LoadViz}
\title{Load data file for ExpoViz module}
\usage{
LoadViz(PID = res$PID, UseExample = "example#1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{UseExample}{chr. Method of uploading data. If "default"，user should upload their own data files,
or use "example#1" provided by this module.}

\item{ExdataPath}{chr. Input data file directory, e.g. "D:/test/eg_expoviz_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VodataPath}{chr. Input vocabulary file directory, e.g. "D:/test/eg_expoviz_voca.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data for visualization.
}
\examples{
res = InitViz()
    res1 = LoadViz(PID = res$PID, UseExample = "example#1")
    res1 =  LoadViz(PID = res$PID,UseExample = "default",
    DataPath = "/Users/lanq/Desktop/test/eg_expoviz_data.xlsx",
    VocaPath = "/Users/lanq/Desktop/test/eg_expoviz_voca.xlsx")
}
\author{
Ning Gao, Bin Wang (corresponding author)
}
