% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponta.R
\name{NtaAnno}
\alias{NtaAnno}
\title{Annotate the non-targeted features}
\usage{
NtaAnno = function(PID, OutPath = "default", VarsY, VarsX  = "default",
    VarsN  = "single.factor", FdrCorrect = F, AdductPos = "all", AdductNeg = "all",Accuracy)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoNTA}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsX}{chr. Exposure variable used for modeling. The default option is
"all.x" (All exposure variables are included). Users can also choose available variables.
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3"}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{FdrCorrect}{lgl. T (or TRUE) and F (or FALSE). Whether to correct the
multiple hypotheses by false positive rate (FDR) method.}

\item{AdductPos}{chr. Adducts formed in the positive mode using LC-MS.
The default is "all", i.e., all the adducts in positive mode will be chosen, including
"M+3ACN+2H,M+ACN+H,M+NH4,M+2ACN+2H,M+2H,M+3H,M+3Na,M+2Na-H,M+ACN+Na,M+H+Na,
M+2K-H,M+H+NH4,2M+K,M+K,2M+NH4,2M+Na,M+2Na,M+DMSO+H,M+2ACN+H,M+IsoProp+Na+H,
M+2H+Na,M+ACN+2H,M+H,2M+H,M+CH3OH+H,M+H+2Na,M+Na,2M+ACN+H,2M+ACN+Na,M+IsoProp+H,M+H+K"}

\item{AdductNeg}{chr. Adducts formed in the negative mode using LC-MS.
The default is "all", i.e., all the adducts in positive mode will be chosen, including
"M+FA-H,M+Hac-H,M+Br,3M-H,2M+Hac-H,M+K-2H,2M+FA-H,M-H,M-H2O-H,M+Na-2H,
M-2H,M+TFA-H,M+Cl,M-3H,2M-H”}

\item{Accuracy}{num. Accuracy threshold to match the target compounds. The default is 1 ppm.}
}
\value{
A list containing non-target analysis results
}
\description{
Annotate the non-targeted features
}
\examples{
res <- InitNTA()
   res = LoadNTA(PID = res$PID, UseExample = "example#1")
   res1 = LoadNTA(PID = res$PID, UseExample = "example#1")
   res2 = DelNearZeroVar(PID = res$PID) #optional, If run, expotidy package is needed
   res3 = DelMiss(PID = res$PID) #optional, If run, expotidy package is needed
   res4 = NtaAnno(PID=res$PID,VarsY = "Y1",VarsX = "default",VarsN = "single.factor",
   FdrCorrect = "F",AdductPos = "M+H",AdductNeg = "M-H",Accuracy = 1)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
