% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expocros.R
\name{VizCrosAsso}
\alias{VizCrosAsso}
\title{Visualize association analysis}
\usage{
VizCrosAsso = function(PID, OutPath = "default",VarsN = "single.factor",
    Layout = "volcano",Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoCros}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "volcano".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
An R6 class object containing the results' plot.
}
\description{
Visualize association analysis
}
\examples{
res <- InitCros()
   res1 = LoadCros(PID = res$PID, UseExample = "example#1")
   res2 = CrosAsso(PID=res$PID, EpiDesign = "cohort", VarsY = "Y1", 
   VarsX = "X5,X6,X7,X8,X9,X10,X11", VarsN = "single.factor" , 
   VarsSel = F, VarsSelThr = 0.1, IncCova = T, Family = "gaussian", 
   RepMsr = F,Corstr = "ar1")
   res13 = VizCrosAsso(PID=res$PID,VarsN="single.factor", Layout = "forest",
   Brightness = "dark",Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang
}
