% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmix.R
\name{VizMixMLR}
\alias{VizMixMLR}
\title{Visualize the model results}
\usage{
VizMixMLR(PID,OutPath = "default",VarsY,SelMethod,Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoMixEffect}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{SelMethod}{chr. Method to select the important features to the final model.
Options include "stepwise" (stepwise regression), "lasso" (Regularization regression of
least absolute shrinkage and selection operator), and "enet" (Regularization regression of elastic net).}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing the MLR analysis plot.
}
\description{
Visualize the results of multiple linear regression (MLR) model
}
\examples{
res <- InitMix()
   res1 = LoadMix(PID = res$PID, UseExample = "example#1")
   res2 = MixMLR(PID = res$PID, VarsY = "Y1", VarsX = "all.x",
   IncCova = "F", SelMethod = "lasso", PredType = "response", Family = "gaussian")
   res3 = VizMixMLR(PID=res$PID, VarsY = "Y1", SelMethod = 'lasso',
   Brightness = "light", Palette = "default1")
}
\author{
Bin Wang
}
