% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exnta.R
\name{VizNtaAnno}
\alias{VizNtaAnno}
\title{Visualize annotation results}
\usage{
VizNtaAnno(PID, OutPath = "default", VarsY, VarsN ,Accuracy = 1,
    Brightness = "light",Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoNTA}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Accuracy}{num. Upper limit of accuracy to match the target molecular. The default is 1.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing plots of non-target annotation analysis results
}
\description{
Visualize annotation results of non-targeted data
}
\examples{
res <- InitNTA()
   res1 = LoadNTA(PID = res$PID, UseExample = "example#1")
   res2 = NtaCros(PID=res$PID, VarsY = "Y1", VarsX = "all.x",
   VarsN = "single.factor", FdrCorrect = T, SelMethod = "all",
   StepwizeThr = 0.1, RF_ImpThr = 0.9, IncCova = "F", Family = "gaussian",
   RepMsr = "F", Corstr = "ar1")
   res3 = NtaAnno(PID=res$PID,VarsY = "Y1",VarsX = "default",VarsN = "single.factor",
   FdrCorrect = "F",AdductPos = "M+H",AdductNeg = "M-H",Accuracy = 1)
   res4 = VizNtaAnno(PID=res$PID, VarsY = "Y1",VarsN = "single.factor",
   Accuracy = 1,Brightness = "light",Palette = "default1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
