% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponta.R
\name{VizNtaCros}
\alias{VizNtaCros}
\title{Visualize the results of Association analysis}
\usage{
VizNtaCros = function(PID, OutPath = "default", VarsY,VarsN, Layout = "volcano",
    EffectThr = 0.5, Brightness = "light", Palette = "default1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by ExpoNTA}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsY}{chr. Outcome variable used for modeling. Only one variable can be entered.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "volcano".}

\item{EffectThr}{num. Threshold of the total importance for the variables to a
random forest model. It ranges 0.5-1.0 with the default value of 0.9.}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list containing plots of non-target analysis association results
}
\description{
Visualize the results of Association analysis for non-targeted data
}
\examples{
res <- InitNTA()
   res = LoadNTA(PID = res$PID, UseExample = "example#1")
   res1 = LoadNTA(PID = res$PID, UseExample = "example#1")
   res2 = DelNearZeroVar(PID = res$PID) #optional, If run, expotidy package is needed
   res3 = DelMiss(PID = res$PID) #optional, If run, expotidy package is needed
   res4 = NtaCros(PID=res$PID, VarsY = "Y2", VarsX = "all.x",
   VarsN = "multiple.factor", FdrCorrect = "T", SelMethod = "lasso",
   StepwizeThr = 0.1, RF_ImpThr = 0.9, IncCova = "F", Family = "binomial",
   RepMsr = "F", Corstr = "ar1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
