% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurv.R
\name{VizSurvAsso}
\alias{VizSurvAsso}
\title{Visualize association analysis}
\usage{
VizSurvAsso = function(PID, OutPath = "default",VarsN = "single.factor",
    Layout = "volcano",Brightness = "light",Palette = "default1",ColorFor= "p.value",SizeFor= "p.value")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitSurv()}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{VarsN}{chr. Choose the single factor or multiple factor model.
Available options include "single.factor" and "multiple.factor"}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "volcano".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., "cell", "nature", "science", "lancet", "nejm", and "jama").}

\item{ColorFor}{chr. Volcano plot dot color. Available options include "p.value" and "hr".}

\item{SizeFor}{chr. Volcano plot dot size. Available options include "p.value" and "hr".}
}
\value{
A list containing the results' plot.
}
\description{
Visualize association analysis
}
\examples{
res <- InitSurv()
   res1 = LoadSurv(PID = res$PID, UseExample = "example#1")
   res2 = TransDummy(PID=res$PID, Vars="default")
   res3 = FindCovaSurv(PID=res$PID, TimeY = "Y1", EventY= 'Y2',
   VarsC_Prior = "default", VarsC_Fixed = NULL, Method = "single.factor", Thr = 0.1)
   res4 = SurvAsso(PID=res$PID, TimeY = "Y1", EventY= 'Y2',VarsX='all.x',
   VarsN="single.factor",VarsSel=T,IncCova=T)
   res5 = VizSurvAsso(PID=res$PID,VarsN="single.factor",Layout="volcano",Brightness= "light",
   Palette = "default1",ColorFor= "p.value",SizeFor= "p.value")
   FuncExit(PID = res$PID)
}
\author{
Changxin Lan, Ning Gao, Bin Wang(corresponding author)
}
